refactor(cli): split tests/cli.rs into per-feature integration tests (#339)
Split the 4380-line tests/cli.rs into 12 feature-scoped integration test
binaries plus a `tests/common/mod.rs` for the shared helpers (fake
toolchain scripts, gc registry seeding, environment harness setup,
etc.). Each new test file is well under the 1000-LOC ceiling.

Refs #339
