--- crates/soldr-cli/tests/cli.rs
@@ -1,4380 +1,0 @@
-use serde_json::Value;
-use std::io::Write;
-use std::process::Command;
-use std::{
-    fs,
-    path::{Path, PathBuf},
-    time::{Duration, Instant, SystemTime, UNIX_EPOCH},
-};
-
-fn rustup_which(tool: &str) -> String {
-    let output = Command::new("rustup")
-        .args(["which", tool])
-        .output()
-        .expect("failed to resolve tool with rustup");
-    assert!(output.status.success(), "rustup which failed for {tool}");
-    String::from_utf8_lossy(&output.stdout).trim().to_string()
-}
-
-fn unique_temp_dir(label: &str) -> PathBuf {
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("time went backwards")
-        .as_nanos();
-    let dir = std::env::temp_dir().join(format!("soldr-{label}-{nanos}"));
-    fs::create_dir_all(&dir).expect("failed to create temp dir");
-    dir
-}
-
-fn toml_string(path: &Path) -> String {
-    path.display()
-        .to_string()
-        .replace('\\', "\\\\")
-        .replace('"', "\\\"")
-}
-
-fn seed_gc_candidate(cache_root: &Path, label: &str) -> PathBuf {
-    let dev_root = cache_root.join("dev-root");
-    let workspace = dev_root.join(label);
-    let target = workspace.join("target");
-    fs::create_dir_all(&target).expect("failed to create target dir");
-    fs::write(target.join("artifact.bin"), b"reclaim me").expect("failed to seed target file");
-    fs::write(
-        cache_root.join("config.toml"),
-        format!("[gc]\nallowlist_roots = [\"{}\"]\n", toml_string(&dev_root)),
-    )
-    .expect("failed to write gc config");
-
-    let registry =
-        soldr_cache::target_registry::TargetRegistry::open(&cache_root.join("state.redb"))
-            .expect("failed to open target registry");
-    let now = soldr_cache::target_registry::current_unix_seconds()
-        .expect("failed to get current unix seconds");
-    registry
-        .upsert_with_time(&target, now - 120)
-        .expect("failed to seed target registry");
-    target
-}
-
-fn seed_gc_file_candidate(cache_root: &Path, label: &str) -> PathBuf {
-    let dev_root = cache_root.join("dev-root");
-    let workspace = dev_root.join(label);
-    fs::create_dir_all(&workspace).expect("failed to create workspace dir");
-    let target = workspace.join("target");
-    fs::write(&target, b"not a directory").expect("failed to seed target file");
-    fs::write(
-        cache_root.join("config.toml"),
-        format!("[gc]\nallowlist_roots = [\"{}\"]\n", toml_string(&dev_root)),
-    )
-    .expect("failed to write gc config");
-
-    let registry =
-        soldr_cache::target_registry::TargetRegistry::open(&cache_root.join("state.redb"))
-            .expect("failed to open target registry");
-    let now = soldr_cache::target_registry::current_unix_seconds()
-        .expect("failed to get current unix seconds");
-    registry
-        .upsert_with_time(&target, now - 120)
-        .expect("failed to seed target registry");
-    target
-}
-
-fn path_display_variants(path: &Path) -> Vec<String> {
-    let mut variants = vec![path.display().to_string()];
-    if let Ok(canonical) = fs::canonicalize(path) {
-        let canonical = canonical.display().to_string();
-        if !variants.contains(&canonical) {
-            variants.push(canonical);
-        }
-    }
-    variants
-}
-
-fn logged_cargo_wrapper(log: &str) -> Option<String> {
-    log.lines().find_map(|line| {
-        let wrapper = line.strip_prefix("cargo wrapper=")?;
-        wrapper
-            .split_once(" rustc=")
-            .map(|(wrapper, _)| wrapper.to_string())
-    })
-}
-
-fn log_contains_owned_soldr_wrapper(log: &str, cache_root: &Path) -> bool {
-    if log.contains(env!("CARGO_BIN_EXE_soldr")) {
-        return true;
-    }
-
-    #[cfg(windows)]
-    {
-        let Some(wrapper) = logged_cargo_wrapper(log) else {
-            return false;
-        };
-        let runtime_root = cache_root.join("runtime").join("soldr-self");
-        path_display_variants(&runtime_root)
-            .iter()
-            .any(|path| wrapper.contains(path))
-    }
-    #[cfg(not(windows))]
-    {
-        let _ = cache_root;
-        false
-    }
-}
-
-fn log_contains_toolchain_homes(
-    log: &str,
-    prefix: &str,
-    cargo_home: &Path,
-    rustup_home: &Path,
-) -> bool {
-    for cargo_home in path_display_variants(cargo_home) {
-        for rustup_home in path_display_variants(rustup_home) {
-            if log.contains(&format!(
-                "{prefix} cargo_home={cargo_home} rustup_home={rustup_home}"
-            )) {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-fn fake_script_path(dir: &Path, name: &str) -> PathBuf {
-    #[cfg(windows)]
-    {
-        dir.join(format!("{name}.cmd"))
-    }
-    #[cfg(not(windows))]
-    {
-        dir.join(name)
-    }
-}
-
-fn write_fake_script(path: &Path, body: &str) {
-    #[cfg(windows)]
-    {
-        fs::write(path, body.replace('\n', "\r\n")).expect("failed to write fake script");
-    }
-    #[cfg(not(windows))]
-    {
-        use std::os::unix::fs::PermissionsExt;
-
-        fs::write(path, body).expect("failed to write fake script");
-        let mut perms = fs::metadata(path)
-            .expect("failed to stat fake script")
-            .permissions();
-        perms.set_mode(0o755);
-        fs::set_permissions(path, perms).expect("failed to chmod fake script");
-    }
-}
-
-fn fake_cargo_script(log_path: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             if \"%~1\"==\"metadata\" (\n\
-               if defined SOLDR_TEST_CARGO_METADATA_PATH (\n\
-                 type \"%SOLDR_TEST_CARGO_METADATA_PATH%\"\n\
-               ) else (\n\
-                 echo cargo wrapper=%RUSTC_WRAPPER% rustc=%RUSTC% cache=%SOLDR_CACHE_ENABLED% session=%ZCCACHE_SESSION_ID% sccache_dir=%SCCACHE_DIR% zccache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-                 for /f \"tokens=1,* delims==\" %%A in ('set CARGO_TARGET_ 2^>nul') do @echo cargo_target_env %%A=%%B>>\"{0}\"\n\
-                 for /f \"tokens=1,* delims==\" %%A in ('set CARGO_PROFILE_ 2^>nul') do @echo cargo_profile_env %%A=%%B>>\"{0}\"\n\
-                 echo {{}}\n\
-               )\n\
-               exit /b 0\n\
-             )\n\
-             if \"%~1\"==\"--version\" (\n\
-               echo cargo version>>\"{0}\"\n\
-               echo cargo 1.0.0-test\n\
-               exit /b 0\n\
-             )\n\
-             echo cargo wrapper=%RUSTC_WRAPPER% rustc=%RUSTC% cache=%SOLDR_CACHE_ENABLED% session=%ZCCACHE_SESSION_ID% sccache_dir=%SCCACHE_DIR% zccache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-             for /f \"tokens=1,* delims==\" %%A in ('set CARGO_TARGET_ 2^>nul') do @echo cargo_target_env %%A=%%B>>\"{0}\"\n\
-             for /f \"tokens=1,* delims==\" %%A in ('set CARGO_PROFILE_ 2^>nul') do @echo cargo_profile_env %%A=%%B>>\"{0}\"\n\
-             if defined RUSTC_WRAPPER (\n\
-             call \"%RUSTC_WRAPPER%\" \"%RUSTC%\" --crate-name demo --emit dep-info,link\n\
-             ) else (\n\
-             call \"%RUSTC%\" --crate-name demo --emit dep-info,link\n\
-             )\n\
-             exit /b %ERRORLEVEL%\n",
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             log_cargo_target_envs() {{\n\
-               env | grep '^CARGO_TARGET_' | while IFS= read -r line; do\n\
-                 echo \"cargo_target_env $line\" >> \"{0}\"\n\
-               done\n\
-             }}\n\
-             log_cargo_profile_envs() {{\n\
-               env | grep '^CARGO_PROFILE_' | while IFS= read -r line; do\n\
-                 echo \"cargo_profile_env $line\" >> \"{0}\"\n\
-               done\n\
-             }}\n\
-             if [ \"$1\" = \"metadata\" ]; then\n\
-               if [ -n \"${{SOLDR_TEST_CARGO_METADATA_PATH:-}}\" ]; then\n\
-                 cat \"$SOLDR_TEST_CARGO_METADATA_PATH\"\n\
-               else\n\
-                 echo \"cargo wrapper=${{RUSTC_WRAPPER:-}} rustc=${{RUSTC:-}} cache=${{SOLDR_CACHE_ENABLED:-}} session=${{ZCCACHE_SESSION_ID:-}} sccache_dir=${{SCCACHE_DIR:-}} zccache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-                 log_cargo_target_envs\n\
-                 log_cargo_profile_envs\n\
-                 echo '{{}}'\n\
-               fi\n\
-               exit 0\n\
-             fi\n\
-             if [ \"$1\" = \"--version\" ]; then\n\
-               echo 'cargo version' >> \"{0}\"\n\
-               echo 'cargo 1.0.0-test'\n\
-               exit 0\n\
-             fi\n\
-             echo \"cargo wrapper=${{RUSTC_WRAPPER:-}} rustc=${{RUSTC:-}} cache=${{SOLDR_CACHE_ENABLED:-}} session=${{ZCCACHE_SESSION_ID:-}} sccache_dir=${{SCCACHE_DIR:-}} zccache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-             log_cargo_target_envs\n\
-             log_cargo_profile_envs\n\
-             if [ -n \"${{RUSTC_WRAPPER:-}}\" ]; then\n\
-               \"$RUSTC_WRAPPER\" \"$RUSTC\" --crate-name demo --emit dep-info,link\n\
-             else\n\
-               \"$RUSTC\" --crate-name demo --emit dep-info,link\n\
-             fi\n",
-            log_path.display()
-        )
-    }
-}
-
-#[cfg(not(windows))]
-fn fake_cargo_with_jobserver_script(log_path: &Path) -> String {
-    format!(
-        "#!/bin/sh\n\
-         echo \"cargo wrapper=${{RUSTC_WRAPPER:-}} rustc=${{RUSTC:-}} cache=${{SOLDR_CACHE_ENABLED:-}} session=${{ZCCACHE_SESSION_ID:-}} zccache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{}\"\n\
-         exec 3</dev/null\n\
-         exec 4>/dev/null\n\
-         export CARGO_MAKEFLAGS='-j --jobserver-fds=3,4'\n\
-         export SOLDR_TEST_JOBSERVER_READ_FD=3\n\
-         export SOLDR_TEST_JOBSERVER_WRITE_FD=4\n\
-         \"$RUSTC_WRAPPER\" \"$RUSTC\" --crate-name demo --emit dep-info,link\n",
-        log_path.display()
-    )
-}
-
-fn fake_rustc_script(log_path: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             if \"%~1\"==\"-Vv\" (\n\
-               echo rustc 1.0.0-test\n\
-               echo host: x86_64-pc-windows-msvc\n\
-               echo release: 1.0.0-test\n\
-               exit /b 0\n\
-             )\n\
-             echo rustc %*>>\"{}\"\n",
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             if [ \"$1\" = \"-Vv\" ]; then\n\
-               echo 'rustc 1.0.0-test'\n\
-               echo 'host: x86_64-unknown-linux-gnu'\n\
-               echo 'release: 1.0.0-test'\n\
-               exit 0\n\
-             fi\n\
-             echo \"rustc $*\" >> \"{}\"\n",
-            log_path.display()
-        )
-    }
-}
-
-fn fake_version_tool_script(log_path: &Path, tool_name: &str) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             echo {0} cargo_home=%CARGO_HOME% rustup_home=%RUSTUP_HOME% args=%*>>\"{1}\"\n\
-             echo {0} 1.0.0 (fake)\n",
-            tool_name,
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             echo \"{0} cargo_home=${{CARGO_HOME:-}} rustup_home=${{RUSTUP_HOME:-}} args=$*\" >> \"{1}\"\n\
-             echo \"{0} 1.0.0 (fake)\"\n",
-            tool_name,
-            log_path.display()
-        )
-    }
-}
-
-fn fake_rustup_script(log_path: &Path, tool_dir: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             echo rustup %* cargo_home=%CARGO_HOME% rustup_home=%RUSTUP_HOME%>>\"{0}\"\n\
-             if \"%~1\"==\"which\" (\n\
-               if \"%~2\"==\"cargo\" (\n\
-                 echo {1}\n\
-                 exit /b 0\n\
-               )\n\
-               if \"%~2\"==\"rustc\" (\n\
-                 echo {2}\n\
-                 exit /b 0\n\
-               )\n\
-               if \"%~2\"==\"rustfmt\" (\n\
-                 echo {3}\n\
-                 exit /b 0\n\
-               )\n\
-             )\n\
-             echo unsupported rustup invocation %* 1>&2\n\
-             exit /b 1\n",
-            log_path.display(),
-            tool_dir.join("cargo.cmd").display(),
-            tool_dir.join("rustc.cmd").display(),
-            tool_dir.join("rustfmt.cmd").display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             echo \"rustup $* cargo_home=${{CARGO_HOME:-}} rustup_home=${{RUSTUP_HOME:-}}\" >> \"{0}\"\n\
-             if [ \"$1\" = \"which\" ]; then\n\
-               case \"$2\" in\n\
-                 cargo)\n\
-                   echo \"{1}\"\n\
-                   exit 0\n\
-                   ;;\n\
-                 rustc)\n\
-                   echo \"{2}\"\n\
-                   exit 0\n\
-                   ;;\n\
-                 rustfmt)\n\
-                   echo \"{3}\"\n\
-                   exit 0\n\
-                   ;;\n\
-               esac\n\
-             fi\n\
-             echo \"unsupported rustup invocation: $*\" >&2\n\
-             exit 1\n",
-            log_path.display(),
-            tool_dir.join("cargo").display(),
-            tool_dir.join("rustc").display(),
-            tool_dir.join("rustfmt").display()
-        )
-    }
-}
-
-fn fake_failing_rustup_script(log_path: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             echo rustup %* cargo_home=%CARGO_HOME% rustup_home=%RUSTUP_HOME%>>\"{}\"\n\
-             echo rustup should not have been invoked 1>&2\n\
-             exit /b 1\n",
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             echo \"rustup $* cargo_home=${{CARGO_HOME:-}} rustup_home=${{RUSTUP_HOME:-}}\" >> \"{}\"\n\
-             echo \"rustup should not have been invoked\" >&2\n\
-             exit 1\n",
-            log_path.display()
-        )
-    }
-}
-
-fn fake_zccache_script(log_path: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             if \"%~1\"==\"start\" goto soldr_zccache_start\n\
-             if \"%~1\"==\"stop\" (\n\
-               echo zccache stop cache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-               if defined SOLDR_TEST_ZCCACHE_STALE_START_ONCE type nul > \"%SOLDR_TEST_ZCCACHE_STALE_START_ONCE%.stopped\"\n\
-               exit /b 0\n\
-             )\n\
-              if \"%~1\"==\"session-start\" (\n\
-                echo zccache session-start cache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-                if not \"%~4\"==\"\" type nul > \"%~4\"\n\
-                if not \"%~6\"==\"\" type nul > \"%~6\"\n\
-                echo {{\"session_id\":\"test-session\"}}\n\
-                exit /b 0\n\
-             )\n\
-             if \"%~1\"==\"session-end\" (\n\
-               echo zccache session-end %~2 %~3 cache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-               if \"%~3\"==\"--json\" (\n\
-                 echo {{\"status\":\"ok\",\"session_id\":\"test-session\",\"duration_ms\":1200,\"compilations\":10,\"hits\":7,\"misses\":3,\"non_cacheable\":2,\"errors\":1,\"time_saved_ms\":900,\"unique_sources\":4,\"bytes_read\":111,\"bytes_written\":222,\"hit_rate\":0.7}}\n\
-               ) else (\n\
-                 echo hits: 1\n\
-               )\n\
-               exit /b 0\n\
-             )\n\
-             if \"%~1\"==\"rust-plan\" (\n\
-               echo zccache rust-plan %~2 cache_dir=%ZCCACHE_CACHE_DIR% args=%*>>\"{0}\"\n\
-               if /I \"%~2\"==\"restore\" if defined SOLDR_TEST_RUST_PLAN_STALE (\n\
-                 echo {{\"operation\":\"restore\",\"restored_file_count\":0,\"artifact_absent_from_restored_plan\":1,\"compatibility\":{{\"status\":\"ok\",\"errors\":[]}}}}\n\
-                 exit /b 0\n\
-               )\n\
-               echo {{\"operation\":\"%~2\",\"compatibility\":{{\"status\":\"ok\",\"errors\":[]}}}}\n\
-               exit /b 0\n\
-             )\n\
-             if \"%~1\"==\"status\" (\n\
-               echo hits=7\n\
-               exit /b 0\n\
-             )\n\
-             if \"%~1\"==\"clear\" (\n\
-               echo zccache clear cache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-               exit /b 0\n\
-             )\n\
-             set \"rustc=%~1\"\n\
-             shift\n\
-             echo zccache wrapper cache_dir=%ZCCACHE_CACHE_DIR% %rustc% %*>>\"{0}\"\n\
-             call \"%rustc%\" %*\n\
-             exit /b %ERRORLEVEL%\n\
-             :soldr_zccache_start\n\
-             echo zccache start cache_dir=%ZCCACHE_CACHE_DIR%>>\"{0}\"\n\
-             if not defined SOLDR_TEST_ZCCACHE_STALE_START_ONCE exit /b 0\n\
-             if exist \"%SOLDR_TEST_ZCCACHE_STALE_START_ONCE%.stopped\" exit /b 0\n\
-             if not exist \"%SOLDR_TEST_ZCCACHE_STALE_START_ONCE%.failed\" (\n\
-               type nul > \"%SOLDR_TEST_ZCCACHE_STALE_START_ONCE%.failed\"\n\
-               echo failed to start daemon: daemon process 3197 exists but not accepting connections 1>&2\n\
-               exit /b 1\n\
-             )\n\
-             echo zccache start retried before stop 1>&2\n\
-             exit /b 66\n",
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             case \"$1\" in\n\
-               start)\n\
-                 echo \"zccache start cache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-                 if [ -n \"${{SOLDR_TEST_ZCCACHE_STALE_START_ONCE:-}}\" ]; then\n\
-                   if [ ! -e \"${{SOLDR_TEST_ZCCACHE_STALE_START_ONCE}}.stopped\" ]; then\n\
-                     if [ ! -e \"${{SOLDR_TEST_ZCCACHE_STALE_START_ONCE}}.failed\" ]; then\n\
-                       : > \"${{SOLDR_TEST_ZCCACHE_STALE_START_ONCE}}.failed\"\n\
-                       echo 'failed to start daemon: daemon process 3197 exists but not accepting connections' >&2\n\
-                       exit 1\n\
-                     fi\n\
-                     echo 'zccache start retried before stop' >&2\n\
-                     exit 66\n\
-                   fi\n\
-                 fi\n\
-                 exit 0\n\
-                 ;;\n\
-               stop)\n\
-                 echo \"zccache stop cache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-                 if [ -n \"${{SOLDR_TEST_ZCCACHE_STALE_START_ONCE:-}}\" ]; then\n\
-                   : > \"${{SOLDR_TEST_ZCCACHE_STALE_START_ONCE}}.stopped\"\n\
-                 fi\n\
-                 exit 0\n\
-                 ;;\n\
-                session-start)\n\
-                  echo \"zccache session-start cache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-                  : > \"$4\"\n\
-                  : > \"$6\"\n\
-                  echo '{{\"session_id\":\"test-session\"}}'\n\
-                  exit 0\n\
-               ;;\n\
-               session-end)\n\
-                 echo \"zccache session-end $2 $3 cache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-                 if [ \"${{3:-}}\" = \"--json\" ]; then\n\
-                   printf '{{\"status\":\"ok\",\"session_id\":\"test-session\",\"duration_ms\":1200,\"compilations\":10,\"hits\":7,\"misses\":3,\"non_cacheable\":2,\"errors\":1,\"time_saved_ms\":900,\"unique_sources\":4,\"bytes_read\":111,\"bytes_written\":222,\"hit_rate\":0.7}}\\n'\n\
-                 else\n\
-                   echo 'hits: 1'\n\
-                 fi\n\
-                 exit 0\n\
-                 ;;\n\
-              rust-plan)\n\
-                echo \"zccache rust-plan $2 cache_dir=${{ZCCACHE_CACHE_DIR:-}} args=$*\" >> \"{0}\"\n\
-                if [ \"$2\" = \"restore\" ] && [ -n \"${{SOLDR_TEST_RUST_PLAN_STALE:-}}\" ]; then\n\
-                  printf '{{\"operation\":\"restore\",\"restored_file_count\":0,\"artifact_absent_from_restored_plan\":1,\"compatibility\":{{\"status\":\"ok\",\"errors\":[]}}}}\\n'\n\
-                  exit 0\n\
-                fi\n\
-                printf '{{\"operation\":\"%s\",\"compatibility\":{{\"status\":\"ok\",\"errors\":[]}}}}\\n' \"$2\"\n\
-                exit 0\n\
-                ;;\n\
-              status)\n\
-                echo 'hits=7'\n\
-                exit 0\n\
-                 ;;\n\
-               clear)\n\
-                 echo \"zccache clear cache_dir=${{ZCCACHE_CACHE_DIR:-}}\" >> \"{0}\"\n\
-                 exit 0\n\
-                 ;;\n\
-             esac\n\
-             rustc=\"$1\"\n\
-             shift\n\
-             if [ -n \"${{SOLDR_TEST_JOBSERVER_READ_FD:-}}\" ]; then\n\
-               if ! eval \": <&$SOLDR_TEST_JOBSERVER_READ_FD\"; then\n\
-                 echo \"jobserver read fd $SOLDR_TEST_JOBSERVER_READ_FD is not open\" >&2\n\
-                 exit 42\n\
-               fi\n\
-               if ! eval \": >&$SOLDR_TEST_JOBSERVER_WRITE_FD\"; then\n\
-                 echo \"jobserver write fd $SOLDR_TEST_JOBSERVER_WRITE_FD is not open\" >&2\n\
-                 exit 42\n\
-               fi\n\
-               echo \"zccache jobserver fds ok read=$SOLDR_TEST_JOBSERVER_READ_FD write=$SOLDR_TEST_JOBSERVER_WRITE_FD\" >> \"{0}\"\n\
-             fi\n\
-             echo \"zccache wrapper cache_dir=${{ZCCACHE_CACHE_DIR:-}} $rustc $*\" >> \"{0}\"\n\
-             \"$rustc\" \"$@\"\n",
-            log_path.display()
-        )
-    }
-}
-
-fn fake_custom_wrapper_script(log_path: &Path, wrapper_name: &str) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             set \"rustc=%~1\"\n\
-             shift\n\
-             echo {1} wrapper %rustc% %*>>\"{0}\"\n\
-             call \"%rustc%\" %*\n\
-             exit /b %ERRORLEVEL%\n",
-            log_path.display(),
-            wrapper_name
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             rustc=\"$1\"\n\
-             shift\n\
-             echo \"{1} wrapper $rustc $*\" >> \"{0}\"\n\
-             \"$rustc\" \"$@\"\n",
-            log_path.display(),
-            wrapper_name
-        )
-    }
-}
-
-fn install_fake_toolchain(log_path: &Path) -> (PathBuf, PathBuf, PathBuf) {
-    let dir = unique_temp_dir("fake-toolchain");
-    let cargo = fake_script_path(&dir, "cargo");
-    let rustc = fake_script_path(&dir, "rustc");
-    let zccache = fake_script_path(&dir, "zccache");
-    write_fake_script(&cargo, &fake_cargo_script(log_path));
-    write_fake_script(&rustc, &fake_rustc_script(log_path));
-    write_fake_script(&zccache, &fake_zccache_script(log_path));
-    (cargo, rustc, zccache)
-}
-
-#[cfg(not(windows))]
-fn install_fake_jobserver_toolchain(log_path: &Path) -> (PathBuf, PathBuf, PathBuf) {
-    let dir = unique_temp_dir("fake-jobserver-toolchain");
-    let cargo = fake_script_path(&dir, "cargo");
-    let rustc = fake_script_path(&dir, "rustc");
-    let zccache = fake_script_path(&dir, "zccache");
-
-    write_fake_script(&cargo, &fake_cargo_with_jobserver_script(log_path));
-    write_fake_script(&rustc, &fake_rustc_script(log_path));
-    write_fake_script(&zccache, &fake_zccache_script(log_path));
-    (cargo, rustc, zccache)
-}
-
-fn install_fake_version_toolchain(tool_dir: &Path, log_path: &Path) -> (PathBuf, PathBuf, PathBuf) {
-    let cargo = fake_script_path(tool_dir, "cargo");
-    let rustc = fake_script_path(tool_dir, "rustc");
-    let rustfmt = fake_script_path(tool_dir, "rustfmt");
-    write_fake_script(&cargo, &fake_version_tool_script(log_path, "cargo"));
-    write_fake_script(&rustc, &fake_version_tool_script(log_path, "rustc"));
-    write_fake_script(&rustfmt, &fake_version_tool_script(log_path, "rustfmt"));
-    (cargo, rustc, rustfmt)
-}
-
-fn install_fake_wrapper(log_path: &Path, wrapper_name: &str) -> PathBuf {
-    let dir = unique_temp_dir("fake-wrapper");
-    let wrapper = fake_script_path(&dir, wrapper_name);
-    write_fake_script(
-        &wrapper,
-        &fake_custom_wrapper_script(log_path, wrapper_name),
-    );
-    wrapper
-}
-
-fn install_fake_rustup_toolchain(log_path: &Path) -> (PathBuf, PathBuf, PathBuf, PathBuf) {
-    let dir = unique_temp_dir("fake-rustup-toolchain");
-    let cargo = fake_script_path(&dir, "cargo");
-    let rustc = fake_script_path(&dir, "rustc");
-    let rustfmt = fake_script_path(&dir, "rustfmt");
-    #[cfg(windows)]
-    let rustup = dir.join("rustup.bat");
-    #[cfg(not(windows))]
-    let rustup = fake_script_path(&dir, "rustup");
-    write_fake_script(&cargo, &fake_version_tool_script(log_path, "cargo"));
-    write_fake_script(&rustc, &fake_version_tool_script(log_path, "rustc"));
-    write_fake_script(&rustfmt, &fake_version_tool_script(log_path, "rustfmt"));
-    write_fake_script(&rustup, &fake_rustup_script(log_path, &dir));
-    (rustup, cargo, rustc, rustfmt)
-}
-
-fn install_failing_fake_rustup(log_path: &Path) -> PathBuf {
-    let dir = unique_temp_dir("fake-rustup-failure");
-    #[cfg(windows)]
-    let rustup = dir.join("rustup.bat");
-    #[cfg(not(windows))]
-    let rustup = fake_script_path(&dir, "rustup");
-    write_fake_script(&rustup, &fake_failing_rustup_script(log_path));
-    rustup
-}
-
-/// Fake rustup that logs one line per invocation to `log_path`, with the
-/// argv joined by `\u{1f}` (ASCII unit separator) so test assertions can
-/// reason about the exact argv even when individual arguments contain
-/// spaces. Always exits 0.
-fn fake_logging_rustup_script(log_path: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             setlocal enabledelayedexpansion\n\
-             set \"line=\"\n\
-             :loop\n\
-             if \"%~1\"==\"\" goto done\n\
-             if defined line (set \"line=!line!\u{1f}%~1\") else (set \"line=%~1\")\n\
-             shift\n\
-             goto loop\n\
-             :done\n\
-             echo !line!>>\"{}\"\n\
-             exit /b 0\n",
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        // Use ASCII Unit Separator (\037) between argv elements so assertions
-        // can split deterministically even if an individual arg contains
-        // spaces. Hand-roll the join in /bin/sh.
-        format!(
-            "#!/bin/sh\n\
-             sep=$(printf '\\037')\n\
-             out=\"\"\n\
-             first=1\n\
-             for arg in \"$@\"; do\n\
-               if [ $first -eq 1 ]; then\n\
-                 out=\"$arg\"\n\
-                 first=0\n\
-               else\n\
-                 out=\"$out${{sep}}$arg\"\n\
-               fi\n\
-             done\n\
-             printf '%s\\n' \"$out\" >> \"{}\"\n\
-             exit 0\n",
-            log_path.display()
-        )
-    }
-}
-
-/// Install a fake rustup that logs argv per invocation. Returns the path
-/// to the fake binary, ready to hand to `SOLDR_TEST_RUSTUP_BIN`.
-fn install_logging_fake_rustup(log_path: &Path) -> PathBuf {
-    let dir = unique_temp_dir("fake-rustup-logging");
-    #[cfg(windows)]
-    let rustup = dir.join("rustup.bat");
-    #[cfg(not(windows))]
-    let rustup = fake_script_path(&dir, "rustup");
-    write_fake_script(&rustup, &fake_logging_rustup_script(log_path));
-    rustup
-}
-
-/// Read every argv invocation logged by `fake_logging_rustup_script`.
-/// Each returned `Vec<String>` is one invocation, with argv split on the
-/// ASCII unit separator.
-fn read_logged_rustup_invocations(log_path: &Path) -> Vec<Vec<String>> {
-    let text = fs::read_to_string(log_path).unwrap_or_default();
-    text.lines()
-        .filter(|line| !line.trim().is_empty())
-        .map(|line| line.split('\u{1f}').map(str::to_string).collect())
-        .collect()
-}
-
-fn seed_rust_toolchain_toml(dir: &Path, contents: &str) {
-    fs::create_dir_all(dir).expect("failed to create workspace dir");
-    fs::write(dir.join("rust-toolchain.toml"), contents)
-        .expect("failed to write rust-toolchain.toml");
-}
-
-/// Fake cargo that logs one line per invocation to `log_path`, with the
-/// argv joined by `\u{1f}` (ASCII unit separator) so test assertions can
-/// reason about the exact argv even when individual arguments contain
-/// spaces. Always exits 0. Mirrors `fake_logging_rustup_script` but
-/// writes to a separate log file so concurrent rustup + cargo
-/// invocations under `toolchain prepare` don't interleave.
-fn fake_logging_cargo_script(log_path: &Path) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             setlocal enabledelayedexpansion\n\
-             set \"line=\"\n\
-             :loop\n\
-             if \"%~1\"==\"\" goto done\n\
-             if defined line (set \"line=!line!\u{1f}%~1\") else (set \"line=%~1\")\n\
-             shift\n\
-             goto loop\n\
-             :done\n\
-             echo !line!>>\"{}\"\n\
-             exit /b 0\n",
-            log_path.display()
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             sep=$(printf '\\037')\n\
-             out=\"\"\n\
-             first=1\n\
-             for arg in \"$@\"; do\n\
-               if [ $first -eq 1 ]; then\n\
-                 out=\"$arg\"\n\
-                 first=0\n\
-               else\n\
-                 out=\"$out${{sep}}$arg\"\n\
-               fi\n\
-             done\n\
-             printf '%s\\n' \"$out\" >> \"{}\"\n\
-             exit 0\n",
-            log_path.display()
-        )
-    }
-}
-
-/// Install a fake cargo that logs argv per invocation. Returns the path
-/// to the fake binary, ready to hand to `SOLDR_TEST_CARGO_BIN`.
-fn install_logging_fake_cargo(log_path: &Path) -> PathBuf {
-    let dir = unique_temp_dir("fake-cargo-logging");
-    let cargo = fake_script_path(&dir, "cargo");
-    write_fake_script(&cargo, &fake_logging_cargo_script(log_path));
-    cargo
-}
-
-/// Read every argv invocation logged by `fake_logging_cargo_script`.
-/// Each returned `Vec<String>` is one invocation, with argv split on
-/// the ASCII unit separator.
-fn read_logged_cargo_invocations(log_path: &Path) -> Vec<Vec<String>> {
-    let text = fs::read_to_string(log_path).unwrap_or_default();
-    text.lines()
-        .filter(|line| !line.trim().is_empty())
-        .map(|line| line.split('\u{1f}').map(str::to_string).collect())
-        .collect()
-}
-
-fn prepend_to_path(dir: &Path) -> std::ffi::OsString {
-    let existing = std::env::var_os("PATH").unwrap_or_default();
-    let mut paths = vec![dir.to_path_buf()];
-    paths.extend(std::env::split_paths(&existing));
-    std::env::join_paths(paths).expect("failed to join PATH")
-}
-
-/// PATH value for tests that need to verify soldr's tool resolution falls back
-/// to its rustup path. Strips the runner's real cargo/rustc entries so
-/// `probe_toolchain_binary`'s PATH search can't shadow the in-test fakes.
-/// On Windows we keep `System32` so `Command::new` can still spawn `.cmd`
-/// shims via `cmd.exe`.
-fn isolated_test_path() -> std::ffi::OsString {
-    #[cfg(windows)]
-    {
-        let system_root = std::env::var_os("SystemRoot")
-            .map(std::path::PathBuf::from)
-            .unwrap_or_else(|| std::path::PathBuf::from(r"C:\Windows"));
-        let dirs = [system_root.join("System32"), system_root];
-        std::env::join_paths(dirs).expect("failed to join isolated PATH")
-    }
-    #[cfg(not(windows))]
-    {
-        std::ffi::OsString::from("/usr/bin:/bin")
-    }
-}
-
-#[test]
-fn version_command_prints_workspace_version() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("version")
-        .output()
-        .expect("failed to run soldr version");
-
-    assert!(output.status.success(), "version command failed");
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert_eq!(
-        stdout.trim(),
-        format!("soldr {}", env!("CARGO_PKG_VERSION"))
-    );
-}
-
-#[test]
-fn version_command_emits_versioned_json() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["version", "--json"])
-        .output()
-        .expect("failed to run soldr version --json");
-
-    assert!(output.status.success(), "version --json command failed");
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("version --json did not return JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "version");
-    assert_eq!(json["soldr_version"], env!("CARGO_PKG_VERSION"));
-}
-
-#[test]
-fn help_lists_phase_one_command_surface() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("--help")
-        .output()
-        .expect("failed to run soldr --help");
-
-    assert!(output.status.success(), "help command failed");
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(stdout.contains("status"), "help output missing status");
-    assert!(stdout.contains("clean"), "help output missing clean");
-    assert!(stdout.contains("purge"), "help output missing purge");
-    assert!(stdout.contains("config"), "help output missing config");
-    assert!(stdout.contains("cache"), "help output missing cache");
-    assert!(stdout.contains("version"), "help output missing version");
-    assert!(stdout.contains("cargo"), "help output missing cargo");
-    assert!(stdout.contains("rustup"), "help output missing rustup");
-    assert!(
-        stdout.contains("toolchain"),
-        "help output missing toolchain"
-    );
-}
-
-#[test]
-fn rustup_passthrough_forwards_args_unchanged_for_unscoped_subcommands() {
-    let workspace = unique_temp_dir("rustup-passthrough-show");
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_logging_fake_rustup(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["rustup", "show"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr rustup show");
-
-    assert!(
-        output.status.success(),
-        "soldr rustup show failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let invocations = read_logged_rustup_invocations(&log_path);
-    assert_eq!(invocations.len(), 1, "expected one rustup invocation");
-    assert_eq!(invocations[0], vec!["show".to_string()]);
-}
-
-#[test]
-fn rustup_passthrough_injects_toolchain_for_target_add() {
-    let workspace = unique_temp_dir("rustup-passthrough-target-add");
-    seed_rust_toolchain_toml(&workspace, "[toolchain]\nchannel = \"1.94.1\"\n");
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_logging_fake_rustup(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["rustup", "target", "add", "x86_64-unknown-linux-musl"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr rustup target add");
-
-    assert!(
-        output.status.success(),
-        "soldr rustup target add failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let invocations = read_logged_rustup_invocations(&log_path);
-    assert_eq!(invocations.len(), 1, "expected one rustup invocation");
-    assert_eq!(
-        invocations[0],
-        vec![
-            "target".to_string(),
-            "add".to_string(),
-            "--toolchain".to_string(),
-            "1.94.1".to_string(),
-            "x86_64-unknown-linux-musl".to_string(),
-        ]
-    );
-}
-
-#[test]
-fn rustup_passthrough_does_not_double_inject_toolchain() {
-    let workspace = unique_temp_dir("rustup-passthrough-explicit-toolchain");
-    seed_rust_toolchain_toml(&workspace, "[toolchain]\nchannel = \"1.94.1\"\n");
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_logging_fake_rustup(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args([
-            "rustup",
-            "target",
-            "add",
-            "--toolchain",
-            "nightly",
-            "aarch64-apple-darwin",
-        ])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr rustup target add --toolchain nightly");
-
-    assert!(
-        output.status.success(),
-        "soldr rustup target add --toolchain failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let invocations = read_logged_rustup_invocations(&log_path);
-    assert_eq!(invocations.len(), 1, "expected one rustup invocation");
-    let invocation = &invocations[0];
-    let toolchain_count = invocation
-        .iter()
-        .filter(|arg| *arg == "--toolchain")
-        .count();
-    assert_eq!(
-        toolchain_count, 1,
-        "--toolchain should appear exactly once: {invocation:?}"
-    );
-    let toolchain_value_idx = invocation
-        .iter()
-        .position(|arg| arg == "--toolchain")
-        .expect("--toolchain not found");
-    assert_eq!(
-        invocation.get(toolchain_value_idx + 1).map(String::as_str),
-        Some("nightly"),
-        "user-supplied toolchain should be preserved: {invocation:?}"
-    );
-}
-
-#[test]
-fn toolchain_install_invokes_rustup_with_channel() {
-    let workspace = unique_temp_dir("toolchain-install");
-    seed_rust_toolchain_toml(&workspace, "[toolchain]\nchannel = \"1.94.1\"\n");
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_logging_fake_rustup(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["toolchain", "install"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr toolchain install");
-
-    assert!(
-        output.status.success(),
-        "soldr toolchain install failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let invocations = read_logged_rustup_invocations(&log_path);
-    assert_eq!(invocations.len(), 1, "expected one rustup invocation");
-    assert_eq!(
-        invocations[0],
-        vec![
-            "toolchain".to_string(),
-            "install".to_string(),
-            "1.94.1".to_string(),
-            "--profile".to_string(),
-            "minimal".to_string(),
-            "--no-self-update".to_string(),
-        ]
-    );
-}
-
-#[test]
-fn toolchain_prepare_installs_channel_components_and_targets() {
-    let workspace = unique_temp_dir("toolchain-prepare");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         components = [\"clippy\"]\n\
-         targets = [\"x86_64-unknown-linux-musl\"]\n",
-    );
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_logging_fake_rustup(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["toolchain", "prepare"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr toolchain prepare");
-
-    assert!(
-        output.status.success(),
-        "soldr toolchain prepare failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let invocations = read_logged_rustup_invocations(&log_path);
-    assert_eq!(
-        invocations.len(),
-        3,
-        "expected install + component add + target add: {invocations:?}"
-    );
-    assert_eq!(
-        invocations[0],
-        vec![
-            "toolchain".to_string(),
-            "install".to_string(),
-            "1.94.1".to_string(),
-            "--profile".to_string(),
-            "minimal".to_string(),
-            "--no-self-update".to_string(),
-        ],
-        "first invocation should install the pinned channel"
-    );
-    assert_eq!(
-        invocations[1],
-        vec![
-            "component".to_string(),
-            "add".to_string(),
-            "--toolchain".to_string(),
-            "1.94.1".to_string(),
-            "clippy".to_string(),
-        ],
-        "second invocation should add the declared component"
-    );
-    assert_eq!(
-        invocations[2],
-        vec![
-            "target".to_string(),
-            "add".to_string(),
-            "--toolchain".to_string(),
-            "1.94.1".to_string(),
-            "x86_64-unknown-linux-musl".to_string(),
-        ],
-        "third invocation should add the declared target"
-    );
-}
-
-#[test]
-fn toolchain_prepare_installs_plugins_with_version() {
-    let workspace = unique_temp_dir("toolchain-prepare-plugin-version");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         \n\
-         [soldr.plugins]\n\
-         cargo-nextest = \"0.9\"\n",
-    );
-    let rustup_log = workspace.join("rustup.log");
-    let cargo_log = workspace.join("cargo.log");
-    let rustup = install_logging_fake_rustup(&rustup_log);
-    let cargo = install_logging_fake_cargo(&cargo_log);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["toolchain", "prepare"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .output()
-        .expect("failed to run soldr toolchain prepare");
-
-    assert!(
-        output.status.success(),
-        "soldr toolchain prepare failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let cargo_invocations = read_logged_cargo_invocations(&cargo_log);
-    assert_eq!(
-        cargo_invocations.len(),
-        1,
-        "expected exactly one cargo install: {cargo_invocations:?}"
-    );
-    let invocation = &cargo_invocations[0];
-    assert_eq!(invocation.first().map(String::as_str), Some("install"));
-    assert_eq!(invocation.get(1).map(String::as_str), Some("cargo-nextest"));
-    let version_idx = invocation
-        .iter()
-        .position(|arg| arg == "--version")
-        .expect("--version should appear");
-    assert_eq!(
-        invocation.get(version_idx + 1).map(String::as_str),
-        Some("0.9")
-    );
-}
-
-#[test]
-fn toolchain_prepare_installs_plugin_with_locked_flag() {
-    let workspace = unique_temp_dir("toolchain-prepare-plugin-locked");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         \n\
-         [soldr.plugins]\n\
-         cargo-zigbuild = { version = \"0.18\", locked = true }\n",
-    );
-    let rustup_log = workspace.join("rustup.log");
-    let cargo_log = workspace.join("cargo.log");
-    let rustup = install_logging_fake_rustup(&rustup_log);
-    let cargo = install_logging_fake_cargo(&cargo_log);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["toolchain", "prepare"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .output()
-        .expect("failed to run soldr toolchain prepare");
-
-    assert!(
-        output.status.success(),
-        "soldr toolchain prepare failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let cargo_invocations = read_logged_cargo_invocations(&cargo_log);
-    assert_eq!(
-        cargo_invocations.len(),
-        1,
-        "expected exactly one cargo install: {cargo_invocations:?}"
-    );
-    let invocation = &cargo_invocations[0];
-    assert_eq!(invocation.first().map(String::as_str), Some("install"));
-    assert_eq!(
-        invocation.get(1).map(String::as_str),
-        Some("cargo-zigbuild")
-    );
-    let version_idx = invocation
-        .iter()
-        .position(|arg| arg == "--version")
-        .expect("--version should appear");
-    assert_eq!(
-        invocation.get(version_idx + 1).map(String::as_str),
-        Some("0.18")
-    );
-    assert!(
-        invocation.iter().any(|arg| arg == "--locked"),
-        "expected --locked in argv: {invocation:?}"
-    );
-}
-
-#[test]
-fn toolchain_prepare_plugin_without_version_uses_no_version_flag() {
-    let workspace = unique_temp_dir("toolchain-prepare-plugin-no-version");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         \n\
-         [soldr.plugins]\n\
-         cargo-deny = \"*\"\n",
-    );
-    let rustup_log = workspace.join("rustup.log");
-    let cargo_log = workspace.join("cargo.log");
-    let rustup = install_logging_fake_rustup(&rustup_log);
-    let cargo = install_logging_fake_cargo(&cargo_log);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["toolchain", "prepare"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .output()
-        .expect("failed to run soldr toolchain prepare");
-
-    assert!(
-        output.status.success(),
-        "soldr toolchain prepare failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let cargo_invocations = read_logged_cargo_invocations(&cargo_log);
-    assert_eq!(
-        cargo_invocations.len(),
-        1,
-        "expected exactly one cargo install: {cargo_invocations:?}"
-    );
-    let invocation = &cargo_invocations[0];
-    assert_eq!(
-        invocation,
-        &vec!["install".to_string(), "cargo-deny".to_string()],
-        "expected bare install argv (no --version for \"*\"): {invocation:?}"
-    );
-    assert!(
-        !invocation.iter().any(|arg| arg == "--version"),
-        "--version should be omitted when spec is \"*\": {invocation:?}"
-    );
-}
-
-#[test]
-fn gc_summary_is_non_destructive_and_lists_largest_candidates() {
-    let cache_root = unique_temp_dir("gc-summary");
-    let target = seed_gc_candidate(&cache_root, "summary-project");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "--older-than", "1s", "--larger-than", "1B"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr gc");
-
-    assert!(
-        output.status.success(),
-        "gc summary failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        target.exists(),
-        "soldr gc summary must not delete {}",
-        target.display()
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("reclaimable:"),
-        "summary should include total reclaimable bytes: {stdout}"
-    );
-    assert!(
-        stdout.contains("largest eligible target directories"),
-        "summary should list largest target directories: {stdout}"
-    );
-    assert!(
-        stdout.contains("Run 'soldr gc purge'"),
-        "summary should point to destructive purge command: {stdout}"
-    );
-}
-
-#[test]
-fn gc_summary_json_reports_candidates_without_deleting() {
-    let cache_root = unique_temp_dir("gc-summary-json");
-    let target = seed_gc_candidate(&cache_root, "summary-json-project");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "--json", "--older-than", "1s", "--larger-than", "1B"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr gc --json");
-
-    assert!(
-        output.status.success(),
-        "gc summary json failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        target.exists(),
-        "soldr gc --json summary must not delete {}",
-        target.display()
-    );
-
-    let json: Value = serde_json::from_slice(&output.stdout).expect("gc --json must be JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "gc");
-    assert_eq!(json["mode"], "summary");
-    assert_eq!(json["dry_run"], true);
-    assert_eq!(json["candidate_count"], 1);
-    assert!(json["total_reclaimable_bytes"].as_u64().unwrap_or(0) > 0);
-    assert_eq!(json["largest_candidates"].as_array().unwrap().len(), 1);
-    assert_eq!(json["deleted_paths"].as_array().unwrap().len(), 0);
-}
-
-#[test]
-fn gc_purge_all_deletes_candidates_without_prompt() {
-    let cache_root = unique_temp_dir("gc-purge-all");
-    let target = seed_gc_candidate(&cache_root, "purge-project");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args([
-            "gc",
-            "purge",
-            "--all",
-            "--older-than",
-            "1s",
-            "--larger-than",
-            "1B",
-        ])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr gc purge --all");
-
-    assert!(
-        output.status.success(),
-        "gc purge --all failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        !target.exists(),
-        "soldr gc purge --all should delete {}",
-        target.display()
-    );
-}
-
-#[test]
-fn gc_purge_enter_accepts_candidate() {
-    let cache_root = unique_temp_dir("gc-purge-enter");
-    let target = seed_gc_candidate(&cache_root, "purge-enter-project");
-
-    let mut child = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "purge", "--older-than", "1s", "--larger-than", "1B"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()
-        .expect("failed to spawn soldr gc purge");
-    child
-        .stdin
-        .as_mut()
-        .expect("stdin should be piped")
-        .write_all(b"\n")
-        .expect("failed to accept prompt");
-    let output = child
-        .wait_with_output()
-        .expect("failed to wait for soldr gc purge");
-
-    assert!(
-        output.status.success(),
-        "gc purge interactive failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        !target.exists(),
-        "Enter should accept and delete {}",
-        target.display()
-    );
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("[Y/n]"),
-        "prompt should advertise default yes: {stderr}"
-    );
-    assert!(
-        stderr.contains("selected 1; succeeded 1; failed 0"),
-        "final summary should include aggregate counts: {stderr}"
-    );
-}
-
-#[test]
-fn gc_purge_all_json_reports_error_log_path_and_keeps_failed_row() {
-    let cache_root = unique_temp_dir("gc-purge-json-failure");
-    let target = seed_gc_file_candidate(&cache_root, "purge-json-failure-project");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args([
-            "gc",
-            "purge",
-            "--all",
-            "--json",
-            "--older-than",
-            "1s",
-            "--larger-than",
-            "1B",
-        ])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr gc purge --all --json");
-
-    assert!(
-        output.status.success(),
-        "gc purge --all --json failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let json: Value = serde_json::from_slice(&output.stdout).expect("gc purge --json must be JSON");
-    assert_eq!(json["mode"], "purge");
-    assert_eq!(json["selected_count"], 1);
-    assert_eq!(json["succeeded_count"], 0);
-    assert_eq!(json["failed_count"], 1);
-    let log_path = PathBuf::from(
-        json["error_log_path"]
-            .as_str()
-            .expect("failure JSON should include error log path"),
-    );
-    assert!(
-        log_path.exists(),
-        "missing gc error log {}",
-        log_path.display()
-    );
-
-    let registry =
-        soldr_cache::target_registry::TargetRegistry::open(&cache_root.join("state.redb"))
-            .expect("failed to open target registry");
-    assert!(
-        registry.get(&target).unwrap().is_some(),
-        "failed deletion row should remain retryable"
-    );
-}
-
-#[test]
-fn gc_list_json_reports_built_project_target_dir() {
-    let cache_root = unique_temp_dir("gc-list-build");
-    let project_dir = unique_temp_dir("gc-list-project");
-
-    fs::write(
-        project_dir.join("Cargo.toml"),
-        "[package]\nname = \"gc_list_demo\"\nversion = \"0.0.1\"\nedition = \"2021\"\n\n[workspace]\n",
-    )
-    .expect("failed to write Cargo.toml");
-    fs::create_dir_all(project_dir.join("src")).expect("failed to create src dir");
-    fs::write(project_dir.join("src/main.rs"), "fn main() {}\n").expect("failed to write main.rs");
-
-    let soldr_bin = env!("CARGO_BIN_EXE_soldr");
-    let cargo = rustup_which("cargo");
-
-    let build = Command::new(&cargo)
-        .args(["build", "--quiet"])
-        .current_dir(&project_dir)
-        .env("RUSTC_WRAPPER", soldr_bin)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        // No zccache binary override → soldr wrapper records the target
-        // dir, then falls through to invoking rustc directly.
-        .env_remove("SOLDR_TEST_ZCCACHE_BIN")
-        .env_remove("ZCCACHE_BINARY")
-        .output()
-        .expect("failed to run cargo build through soldr wrapper");
-
-    assert!(
-        build.status.success(),
-        "cargo build failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&build.stdout),
-        String::from_utf8_lossy(&build.stderr)
-    );
-
-    let target_dir = project_dir.join("target");
-    assert!(target_dir.exists(), "target/ should exist after build");
-
-    let canonical_target = fs::canonicalize(&target_dir).unwrap_or_else(|_| target_dir.clone());
-
-    let output = Command::new(soldr_bin)
-        .args(["gc", "list", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr gc list --json");
-
-    assert!(
-        output.status.success(),
-        "gc list --json failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value = serde_json::from_slice(&output.stdout).expect("gc list --json must be JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "gc");
-    assert_eq!(json["mode"], "list");
-    let entry_count = json["entry_count"].as_u64().expect("entry_count");
-    assert!(entry_count >= 1, "expected at least one tracked target dir");
-
-    let entries = json["entries"].as_array().expect("entries array");
-    let canonical_str = canonical_target.display().to_string();
-    let target_str = target_dir.display().to_string();
-    let matched = entries.iter().find(|e| {
-        let p = e["path"].as_str().unwrap_or("");
-        let pb = PathBuf::from(p);
-        if p == canonical_str || p == target_str || pb == canonical_target || pb == target_dir {
-            return true;
-        }
-        if let Ok(canon_p) = fs::canonicalize(&pb) {
-            return canon_p == canonical_target;
-        }
-        false
-    });
-    let entry = matched.unwrap_or_else(|| {
-        panic!(
-            "target dir {} not found in gc list entries: {}",
-            canonical_target.display(),
-            serde_json::to_string_pretty(entries).unwrap()
-        )
-    });
-
-    let path_str = entry["path"].as_str().expect("entry path");
-    assert!(
-        PathBuf::from(path_str).is_absolute(),
-        "gc list entries must use absolute paths: {path_str}"
-    );
-    let size_bytes = entry["size_bytes"].as_u64().expect("size_bytes");
-    assert!(size_bytes > 0, "built target/ should have non-zero size");
-    let file_count = entry["file_count"].as_u64().expect("file_count");
-    assert!(file_count > 0, "built target/ should contain files");
-
-    for entry in entries {
-        assert!(
-            entry["path"].is_string(),
-            "every entry must have a path string"
-        );
-        assert!(
-            entry["size_bytes"].is_u64(),
-            "every entry must have size_bytes"
-        );
-        assert!(
-            entry["file_count"].is_u64(),
-            "every entry must have file_count"
-        );
-        assert!(
-            entry.get("exists").is_none(),
-            "missing rows are pruned, so `exists` should not appear on listed entries"
-        );
-    }
-}
-
-#[test]
-fn gc_list_json_prunes_missing_registry_rows_in_one_pass() {
-    let cache_root = unique_temp_dir("gc-list-prune");
-    let dev_root = cache_root.join("dev-root");
-
-    let live_workspace = dev_root.join("live-project");
-    let live_target = live_workspace.join("target");
-    fs::create_dir_all(&live_target).expect("failed to create live target dir");
-    fs::write(live_target.join("artifact.bin"), b"keep me").expect("failed to seed live target");
-
-    let missing_target = dev_root.join("ghost-project").join("target");
-
-    {
-        let registry =
-            soldr_cache::target_registry::TargetRegistry::open(&cache_root.join("state.redb"))
-                .expect("failed to open target registry");
-        let now = soldr_cache::target_registry::current_unix_seconds()
-            .expect("failed to read clock for seeding");
-        registry
-            .upsert_with_time(&live_target, now - 30)
-            .expect("failed to seed live row");
-        registry
-            .upsert_with_time(&missing_target, now - 30)
-            .expect("failed to seed missing row");
-        assert!(
-            registry.get(&missing_target).unwrap().is_some(),
-            "missing-row seed precondition"
-        );
-    }
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "list", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr gc list --json");
-    assert!(
-        output.status.success(),
-        "gc list --json failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value = serde_json::from_slice(&output.stdout).expect("gc list must be JSON");
-    assert_eq!(json["mode"], "list");
-    assert_eq!(json["entry_count"], 1);
-    assert_eq!(json["pruned_missing"], 1);
-
-    let entries = json["entries"].as_array().expect("entries array");
-    assert_eq!(entries.len(), 1, "missing entry must not be reported");
-    let missing_str = missing_target.display().to_string();
-    for entry in entries {
-        let p = entry["path"].as_str().unwrap_or("");
-        assert_ne!(p, missing_str, "missing path leaked into output");
-    }
-
-    let registry_after =
-        soldr_cache::target_registry::TargetRegistry::open(&cache_root.join("state.redb"))
-            .expect("failed to reopen registry");
-    assert!(
-        registry_after.get(&missing_target).unwrap().is_none(),
-        "missing row should be batched out of the registry"
-    );
-    assert!(
-        registry_after.get(&live_target).unwrap().is_some(),
-        "live row must be preserved"
-    );
-}
-
-#[test]
-fn gc_flat_all_is_rejected_with_purge_hint() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "--all"])
-        .env("SOLDR_CACHE_DIR", unique_temp_dir("gc-flat-all"))
-        .output()
-        .expect("failed to run soldr gc --all");
-
-    assert!(
-        !output.status.success(),
-        "legacy flat gc --all must not silently delete"
-    );
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("soldr gc purge --all"),
-        "error should point to the new purge command: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_front_door_runs_real_cargo() {
-    let cache_root = unique_temp_dir("cargo-version");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "--version"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr cargo --version");
-
-    assert!(output.status.success(), "cargo front door failed");
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stdout.contains("cargo"),
-        "unexpected cargo output: {stdout}"
-    );
-    assert!(
-        !stderr.contains("soldr: fetching cargo"),
-        "cargo front door should not fetch cargo: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_front_door_consumes_no_cache_flag() {
-    let cache_root = unique_temp_dir("cargo-no-cache");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "--version"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr --no-cache cargo --version");
-
-    assert!(
-        output.status.success(),
-        "cargo front door with top-level --no-cache failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stdout.contains("cargo"),
-        "unexpected cargo output with --no-cache: {stdout}"
-    );
-    assert!(
-        !stderr.contains("unexpected argument '--no-cache'"),
-        "--no-cache should be consumed by soldr, not forwarded to cargo: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_build_warns_when_disk_space_is_low() {
-    let cache_root = unique_temp_dir("cargo-low-disk");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_TEST_FREE_DISK_BYTES", "1500000000")
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with low-disk override");
-
-    assert!(
-        output.status.success(),
-        "cargo build with low-disk warning failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("disk space is low"),
-        "low-disk warning missing from stderr: {stderr}"
-    );
-    assert!(
-        stderr.contains("Run `soldr gc`"),
-        "low-disk warning should recommend soldr gc: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_build_ignores_disk_space_detection_failures() {
-    let cache_root = unique_temp_dir("cargo-low-disk-error");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_TEST_FREE_DISK_BYTES", "error")
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with disk-probe error");
-
-    assert!(
-        output.status.success(),
-        "disk-space detection failure must not fail build\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("disk space is low"),
-        "disk-space probe failure should not emit low-disk warning: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_subcommand_rejects_no_cache_flag() {
-    let cache_root = unique_temp_dir("cargo-subcommand-no-cache");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "--no-cache", "--version"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr cargo --no-cache --version");
-
-    assert!(
-        !output.status.success(),
-        "cargo subcommand form should no longer accept --no-cache"
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("--no-cache"),
-        "expected cargo-subcommand form to fail mentioning --no-cache: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_front_door_uses_soldr_wrapper_and_managed_zccache_by_default() {
-    let cache_root = unique_temp_dir("cargo-default-cache");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    // CI (setup-soldr) exports `SOLDR_TARGET_CACHE_MODE=full` /
-    // `SOLDR_BUILD_CACHE_MODE=full` in the job environment. Without
-    // stripping them, the spawned soldr triggers its rust-plan path,
-    // shells out to the fake cargo for `cargo metadata`, gets back `{}`,
-    // and fails with "missing field `packages`". Every fake-toolchain
-    // test below clears these the same way for the same reason.
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with fake zccache");
-
-    assert!(
-        output.status.success(),
-        "cache-enabled front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cargo wrapper="),
-        "fake cargo did not record wrapper env: {log}"
-    );
-    assert!(
-        log_contains_owned_soldr_wrapper(&log, &cache_root),
-        "soldr should own the wrapper slot in cache-enabled mode: {log}"
-    );
-    assert!(
-        log.contains("cache=1"),
-        "cache-enabled front door should propagate cache flag: {log}"
-    );
-    let zccache_cache_dir = cache_root.join("cache").join("zccache");
-    assert!(
-        path_display_variants(&zccache_cache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("zccache_dir={path}"))
-                && log.contains(&format!("cache_dir={path}"))),
-        "managed zccache commands and cargo wrapper env should use the Soldr-owned cache dir: {log}"
-    );
-    assert!(
-        log.contains("zccache start"),
-        "managed zccache should be started for cache-enabled builds: {log}"
-    );
-    assert!(
-        log.contains("zccache session-start"),
-        "managed zccache session should start before cargo runs: {log}"
-    );
-    assert!(
-        log.contains("zccache wrapper"),
-        "wrapper mode should dispatch into zccache on cache-enabled builds: {log}"
-    );
-    assert!(
-        log.contains("rustc ") && log.contains("--crate-name demo"),
-        "real rustc invocation should still happen through the wrapper path: {log}"
-    );
-    assert!(
-        log.contains("zccache session-end test-session --json"),
-        "managed zccache session should end after cargo finishes: {log}"
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("soldr: zccache session summary"),
-        "expected zccache session summary in stderr: {stderr}"
-    );
-    assert!(
-        stderr.contains("hits: 7") && stderr.contains("misses: 3"),
-        "expected zccache hit/miss counts in stderr: {stderr}"
-    );
-
-    let journal = zccache_cache_dir.join("logs").join("last-session.jsonl");
-    let session_log = zccache_cache_dir.join("logs").join("last-session.log");
-    let session_stats = zccache_cache_dir
-        .join("logs")
-        .join("last-session-stats.json");
-    assert!(
-        session_log.exists(),
-        "expected session log at {}",
-        session_log.display()
-    );
-    assert!(
-        journal.exists(),
-        "expected session journal at {}",
-        journal.display()
-    );
-    let stats_json: serde_json::Value = serde_json::from_str(
-        &fs::read_to_string(&session_stats)
-            .unwrap_or_else(|err| panic!("failed to read {}: {err}", session_stats.display())),
-    )
-    .expect("failed to parse zccache session stats JSON");
-    assert_eq!(stats_json["status"], "ok");
-    assert_eq!(stats_json["hits"], 7);
-    assert_eq!(stats_json["misses"], 3);
-}
-
-#[cfg(windows)]
-#[test]
-fn windows_worktree_copy_relocates_wrapper_and_original_dir_can_be_removed() {
-    let cache_root = unique_temp_dir("windows-self-relocate-cache");
-    let worktree = unique_temp_dir("windows-self-relocate-worktree");
-    let source_dir = worktree.join("target").join("debug");
-    fs::create_dir_all(&source_dir).expect("failed to create copied exe dir");
-    let copied_soldr = source_dir.join("soldr.exe");
-    fs::copy(env!("CARGO_BIN_EXE_soldr"), &copied_soldr)
-        .expect("failed to copy soldr exe into temporary worktree");
-
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(&copied_soldr)
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run copied soldr exe");
-
-    assert!(
-        output.status.success(),
-        "copied soldr front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    let wrapper = logged_cargo_wrapper(&log).expect("fake cargo should log RUSTC_WRAPPER");
-    assert!(
-        path_display_variants(&cache_root.join("runtime").join("soldr-self"))
-            .iter()
-            .any(|path| wrapper.contains(path)),
-        "RUSTC_WRAPPER should point at the relocated runtime copy: {log}"
-    );
-    assert!(
-        !path_display_variants(&copied_soldr)
-            .iter()
-            .any(|path| wrapper.contains(path)),
-        "RUSTC_WRAPPER should not point at the original worktree copy: {log}"
-    );
-
-    fs::remove_dir_all(&worktree)
-        .expect("temporary worktree should be removable after soldr exits");
-    assert!(!worktree.exists());
-}
-
-#[test]
-fn cargo_front_door_defaults_dev_debug_off_and_warns_once_per_repo() {
-    let cache_root = unique_temp_dir("cargo-debug-default-off");
-    let repo = unique_temp_dir("cargo-debug-default-repo");
-    fs::write(
-        repo.join("Cargo.toml"),
-        "[package]\nname = \"demo\"\nversion = \"0.1.0\"\nedition = \"2021\"\n",
-    )
-    .expect("failed to seed Cargo.toml");
-    fs::create_dir_all(repo.join("src")).expect("failed to create src dir");
-    fs::write(repo.join("src").join("lib.rs"), "").expect("failed to seed lib.rs");
-
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let cargo_home = cache_root.join("cargo-home");
-
-    let first = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .current_dir(&repo)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("CARGO_HOME", &cargo_home)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("CARGO_PROFILE_DEV_DEBUG")
-        .env_remove("CARGO_PROFILE_TEST_DEBUG")
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run first soldr cargo build");
-    assert!(
-        first.status.success(),
-        "first build failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&first.stdout),
-        String::from_utf8_lossy(&first.stderr)
-    );
-
-    let second = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .current_dir(&repo)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("CARGO_HOME", &cargo_home)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("CARGO_PROFILE_DEV_DEBUG")
-        .env_remove("CARGO_PROFILE_TEST_DEBUG")
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run second soldr cargo build");
-    assert!(
-        second.status.success(),
-        "second build failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&second.stdout),
-        String::from_utf8_lossy(&second.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cargo_profile_env CARGO_PROFILE_DEV_DEBUG=false"),
-        "soldr should inject the dev profile debug override when unspecified: {log}"
-    );
-
-    let first_stderr = String::from_utf8_lossy(&first.stderr);
-    assert!(
-        first_stderr.contains("Cargo profile.dev.debug is unspecified")
-            && first_stderr.contains("CARGO_PROFILE_DEV_DEBUG=false"),
-        "first invocation should warn about the defaulted debug setting: {first_stderr}"
-    );
-    let second_stderr = String::from_utf8_lossy(&second.stderr);
-    assert!(
-        !second_stderr.contains("Cargo profile.dev.debug is unspecified"),
-        "second invocation for the same repo should not repeat the debug-default warning: {second_stderr}"
-    );
-}
-
-#[test]
-fn cargo_front_door_respects_dev_debug_in_cargo_config_toml() {
-    let cache_root = unique_temp_dir("cargo-debug-config-explicit");
-    let repo = unique_temp_dir("cargo-debug-config-repo");
-    fs::write(
-        repo.join("Cargo.toml"),
-        "[package]\nname = \"demo\"\nversion = \"0.1.0\"\nedition = \"2021\"\n",
-    )
-    .expect("failed to seed Cargo.toml");
-    fs::create_dir_all(repo.join(".cargo")).expect("failed to create .cargo dir");
-    fs::write(
-        repo.join(".cargo").join("config.toml"),
-        "[profile.dev]\ndebug = true\n",
-    )
-    .expect("failed to seed .cargo/config.toml");
-
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .current_dir(&repo)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("CARGO_HOME", cache_root.join("cargo-home"))
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("CARGO_PROFILE_DEV_DEBUG")
-        .env_remove("CARGO_PROFILE_TEST_DEBUG")
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with explicit config debug");
-
-    assert!(
-        output.status.success(),
-        "build with explicit cargo config debug failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        !log.contains("cargo_profile_env CARGO_PROFILE_DEV_DEBUG=false"),
-        "explicit .cargo/config.toml profile.dev.debug must not be overridden: {log}"
-    );
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("Cargo profile.dev.debug is unspecified"),
-        "explicit .cargo/config.toml profile.dev.debug should suppress warning: {stderr}"
-    );
-}
-
-/// Detect lines of the form `cargo_target_env CARGO_TARGET_<TRIPLE>_LINKER=<value>`
-/// emitted by the fake cargo script. Returns the `<value>` of the first match.
-fn extract_linker_env_value(log: &str) -> Option<String> {
-    extract_cargo_target_env_value(log, "LINKER")
-}
-
-fn extract_rustflags_env_value(log: &str) -> Option<String> {
-    extract_cargo_target_env_value(log, "RUSTFLAGS")
-}
-
-fn extract_cargo_target_env_value(log: &str, suffix: &str) -> Option<String> {
-    for line in log.lines() {
-        let Some(rest) = line.strip_prefix("cargo_target_env ") else {
-            continue;
-        };
-        let Some(eq_idx) = rest.find('=') else {
-            continue;
-        };
-        let (name, value) = (&rest[..eq_idx], &rest[eq_idx + 1..]);
-        if name.starts_with("CARGO_TARGET_") && name.ends_with(&format!("_{suffix}")) {
-            return Some(value.trim().to_string());
-        }
-    }
-    None
-}
-
-fn log_has_any_cargo_target_env(log: &str) -> bool {
-    log.lines()
-        .any(|line| line.starts_with("cargo_target_env "))
-}
-
-#[test]
-fn cargo_front_door_default_linker_does_not_inject_target_env() {
-    let cache_root = unique_temp_dir("cargo-default-linker");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .env_remove("SOLDR_LINKER")
-        .output()
-        .expect("failed to run soldr cargo build with no SOLDR_LINKER");
-
-    assert!(
-        output.status.success(),
-        "default-linker front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        !log_has_any_cargo_target_env(&log),
-        "default linker should not inject any CARGO_TARGET_* env: {log}"
-    );
-}
-
-#[test]
-fn cargo_front_door_rust_lld_injects_target_linker_env() {
-    let cache_root = unique_temp_dir("cargo-rust-lld-linker");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .env("SOLDR_LINKER", "rust-lld")
-        .output()
-        .expect("failed to run soldr cargo build with SOLDR_LINKER=rust-lld");
-
-    assert!(
-        output.status.success(),
-        "rust-lld front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    let linker_value = extract_linker_env_value(&log).unwrap_or_else(|| {
-        panic!("expected CARGO_TARGET_<triple>_LINKER in fake cargo log: {log}")
-    });
-    let rustflags_value = extract_rustflags_env_value(&log);
-
-    #[cfg(target_os = "windows")]
-    {
-        assert_eq!(
-            linker_value, "rust-lld",
-            "windows-msvc rust-lld should inject rust-lld directly: {log}"
-        );
-        assert!(
-            rustflags_value.is_none(),
-            "windows-msvc rust-lld should not inject rustflags: {log}"
-        );
-    }
-    #[cfg(not(target_os = "windows"))]
-    {
-        assert_eq!(
-            linker_value, "clang",
-            "non-windows rust-lld should drive linking through clang: {log}"
-        );
-        assert_eq!(
-            rustflags_value.as_deref(),
-            Some("-C link-arg=-fuse-ld=lld"),
-            "non-windows rust-lld should add -fuse-ld=lld rustflag: {log}"
-        );
-    }
-}
-
-#[cfg(any(windows, target_os = "macos"))]
-#[test]
-fn cargo_front_door_mold_on_non_linux_returns_clear_error() {
-    let cache_root = unique_temp_dir("cargo-mold-non-linux");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .env("SOLDR_LINKER", "mold")
-        .output()
-        .expect("failed to run soldr cargo build with SOLDR_LINKER=mold on non-linux");
-
-    assert!(
-        !output.status.success(),
-        "SOLDR_LINKER=mold should fail on non-linux hosts; stdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("mold is not supported"),
-        "non-linux mold error message should mention 'mold is not supported': {stderr}"
-    );
-}
-
-/// `SOLDR_LINKER=fast` picks `rust-lld` on every non-Linux host. The
-/// Linux variant of this matrix is exercised by the unit tests in
-/// `crates/soldr-cli/src/linker.rs` (the `mold_present` probe is split
-/// out for testability there). Gating to non-Linux here keeps the
-/// integration test from depending on whether mold happens to be on
-/// `PATH` on the CI runner.
-#[cfg(not(target_os = "linux"))]
-#[test]
-fn cargo_front_door_fast_picks_rust_lld_when_mold_absent() {
-    let cache_root = unique_temp_dir("cargo-fast-linker");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .env("SOLDR_LINKER", "fast")
-        .output()
-        .expect("failed to run soldr cargo build with SOLDR_LINKER=fast");
-
-    assert!(
-        output.status.success(),
-        "fast-linker front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    let linker_value = extract_linker_env_value(&log).unwrap_or_else(|| {
-        panic!("expected CARGO_TARGET_<triple>_LINKER in fake cargo log: {log}")
-    });
-
-    #[cfg(target_os = "windows")]
-    {
-        assert_eq!(
-            linker_value, "rust-lld",
-            "windows-msvc fast should inject rust-lld directly: {log}"
-        );
-    }
-    #[cfg(target_os = "macos")]
-    {
-        assert_eq!(
-            linker_value, "clang",
-            "macOS fast should drive linking through clang: {log}"
-        );
-        assert_eq!(
-            extract_rustflags_env_value(&log).as_deref(),
-            Some("-C link-arg=-fuse-ld=lld"),
-            "macOS fast should add -fuse-ld=lld rustflag: {log}"
-        );
-    }
-}
-
-#[test]
-fn cargo_front_door_invokes_zccache_rust_plan_when_target_cache_enabled() {
-    let cache_root = unique_temp_dir("cargo-rust-plan-cache");
-    let workspace = unique_temp_dir("cargo-rust-plan-workspace");
-    let plan_cache = cache_root.join("target-artifact-cache");
-    let zccache_cache_dir = cache_root.join("cache").join("zccache");
-    let log_path = cache_root.join("tool.log");
-    let metadata_path = cache_root.join("metadata.json");
-    let target_dir = workspace.join("target");
-    fs::create_dir_all(workspace.join("app/src")).expect("create app source");
-    fs::create_dir_all(&target_dir).expect("create target dir");
-    fs::write(workspace.join("Cargo.lock"), "# lock\n").expect("write lockfile");
-    fs::write(workspace.join("Cargo.toml"), "[workspace]\n").expect("write workspace manifest");
-    fs::write(workspace.join("app/Cargo.toml"), "[package]\nname='app'\n")
-        .expect("write app manifest");
-
-    let metadata = serde_json::json!({
-        "packages": [
-            {
-                "id": "path+file:///repo/app#app@0.1.0",
-                "source": null
-            },
-            {
-                "id": "registry+https://github.com/rust-lang/crates.io-index#serde@1.0.0",
-                "source": "registry+https://github.com/rust-lang/crates.io-index"
-            }
-        ],
-        "workspace_members": ["path+file:///repo/app#app@0.1.0"],
-        "workspace_root": workspace,
-        "target_directory": target_dir
-    });
-    fs::write(&metadata_path, serde_json::to_string(&metadata).unwrap())
-        .expect("write metadata fixture");
-
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build", "--locked"])
-        .current_dir(&workspace)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_TEST_CARGO_METADATA_PATH", &metadata_path)
-        .env("SOLDR_TARGET_CACHE_MODE", "thin")
-        .env("SOLDR_TARGET_CACHE_BUNDLE_DIR", &plan_cache)
-        .output()
-        .expect("failed to run soldr cargo build with rust-plan target cache");
-
-    assert!(
-        output.status.success(),
-        "cache-enabled rust-plan front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("zccache rust-plan restore") && log.contains("zccache rust-plan save"),
-        "soldr should call zccache rust-plan restore/save when target cache is enabled: {log}"
-    );
-    assert!(
-        path_display_variants(&zccache_cache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("cache_dir={path}"))),
-        "rust-plan calls should query the active managed zccache daemon cache dir: {log}"
-    );
-    assert!(
-        path_display_variants(&plan_cache).iter().any(|path| {
-            log.contains(&format!("--cache-dir {path}"))
-                || log.contains(&format!("--cache-dir \"{path}\""))
-        }),
-        "rust-plan calls should still pass the target artifact bundle path via --cache-dir: {log}"
-    );
-    assert!(
-        log.find("zccache rust-plan restore") < log.find("cargo wrapper=")
-            && log.find("cargo wrapper=") < log.find("zccache rust-plan save"),
-        "rust-plan restore should run before Cargo and save should run after Cargo: {log}"
-    );
-
-    let plan_path = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("plans")
-        .join("last-rust-artifact-plan.json");
-    let plan: Value =
-        serde_json::from_str(&fs::read_to_string(&plan_path).expect("read generated rust plan"))
-            .expect("parse generated rust plan");
-    assert_eq!(plan["schema_version"], 1);
-    assert_eq!(plan["mode"], "thin");
-    assert_eq!(plan["cache_schema_version"], 1);
-    assert_eq!(
-        plan["packages"]["workspace_package_ids"][0],
-        "path+file:///repo/app#app@0.1.0"
-    );
-    assert!(
-        plan["packages"]["selected_package_ids"][0]
-            .as_str()
-            .unwrap()
-            .contains("serde"),
-        "external dependency should be selected in generated plan: {plan}"
-    );
-}
-
-#[test]
-fn cargo_front_door_warns_when_rust_plan_restore_is_partial() {
-    let cache_root = unique_temp_dir("cargo-rust-plan-partial-cache");
-    let workspace = unique_temp_dir("cargo-rust-plan-partial-workspace");
-    let plan_cache = cache_root.join("target-artifact-cache");
-    let log_path = cache_root.join("tool.log");
-    let metadata_path = cache_root.join("metadata.json");
-    let target_dir = workspace.join("target");
-    fs::create_dir_all(workspace.join("app/src")).expect("create app source");
-    fs::create_dir_all(&target_dir).expect("create target dir");
-    fs::write(workspace.join("Cargo.lock"), "# lock\n").expect("write lockfile");
-    fs::write(workspace.join("Cargo.toml"), "[workspace]\n").expect("write workspace manifest");
-    fs::write(workspace.join("app/Cargo.toml"), "[package]\nname='app'\n")
-        .expect("write app manifest");
-
-    let metadata = serde_json::json!({
-        "packages": [
-            {
-                "id": "path+file:///repo/app#app@0.1.0",
-                "source": null
-            }
-        ],
-        "workspace_members": ["path+file:///repo/app#app@0.1.0"],
-        "workspace_root": workspace,
-        "target_directory": target_dir
-    });
-    fs::write(&metadata_path, serde_json::to_string(&metadata).unwrap())
-        .expect("write metadata fixture");
-
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build", "--locked"])
-        .current_dir(&workspace)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_TEST_CARGO_METADATA_PATH", &metadata_path)
-        .env("SOLDR_TARGET_CACHE_MODE", "thin")
-        .env("SOLDR_TARGET_CACHE_BUNDLE_DIR", &plan_cache)
-        .env("SOLDR_TEST_RUST_PLAN_STALE", "1")
-        .output()
-        .expect("failed to run soldr cargo build with stale rust-plan restore");
-
-    assert!(
-        output.status.success(),
-        "soldr should continue after a partial rust-plan restore so Cargo can still run\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("rust-plan restore is partial"),
-        "soldr should warn when zccache reports artifact_absent_from_restored_plan > 0: {stderr}"
-    );
-    assert!(
-        stderr.contains("target-dir") || stderr.contains("--target-dir"),
-        "warning should point users at the shared target-dir workaround: {stderr}"
-    );
-    assert!(
-        stderr.contains("228"),
-        "warning should reference issue #228 so users can find context: {stderr}"
-    );
-}
-
-#[test]
-fn cargo_front_door_recovers_from_stale_zccache_daemon_start() {
-    let cache_root = unique_temp_dir("cargo-stale-zccache-daemon");
-    let log_path = cache_root.join("tool.log");
-    let stale_marker = cache_root.join("stale-zccache-start");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_TEST_ZCCACHE_STALE_START_ONCE", &stale_marker)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with stale fake zccache");
-
-    assert!(
-        output.status.success(),
-        "cache-enabled front door should recover from stale zccache daemon\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert_eq!(
-        log.matches("zccache start").count(),
-        2,
-        "managed zccache start should be retried once after stale daemon detection: {log}"
-    );
-    assert!(
-        log.contains("zccache stop"),
-        "managed zccache should stop stale daemon state before retrying: {log}"
-    );
-    assert!(
-        log.contains("zccache session-start")
-            && log.contains("zccache wrapper")
-            && log.contains("zccache session-end test-session"),
-        "recovered build should still exercise the normal managed zccache path: {log}"
-    );
-    assert!(
-        stale_marker.with_extension("stopped").exists(),
-        "fake zccache should only allow recovery after stop"
-    );
-}
-
-#[test]
-fn cargo_front_door_uses_real_tool_overrides_before_path_probe() {
-    let cache_root = unique_temp_dir("cargo-real-tool-overrides");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let shim_dir = unique_temp_dir("cargo-shim-dir");
-    let shim_cargo = fake_script_path(&shim_dir, "cargo");
-    write_fake_script(
-        &shim_cargo,
-        &fake_version_tool_script(&log_path, "shim-cargo"),
-    );
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_REAL_CARGO", &cargo)
-        .env("SOLDR_REAL_RUSTC", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("PATH", prepend_to_path(&shim_dir))
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with real tool overrides");
-
-    assert!(
-        output.status.success(),
-        "real-tool override front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cargo wrapper="),
-        "real cargo should have been invoked: {log}"
-    );
-    assert!(
-        !log.contains("shim-cargo"),
-        "PATH shim should not be resolved when SOLDR_REAL_CARGO is set: {log}"
-    );
-}
-
-#[test]
-fn cargo_front_door_does_not_start_cache_for_non_build_subcommands() {
-    let cache_root = unique_temp_dir("cargo-non-build-no-cache");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "metadata"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .output()
-        .expect("failed to run soldr cargo metadata with fake tools");
-
-    assert!(
-        output.status.success(),
-        "non-build cargo front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cache=0"),
-        "non-build cargo commands should propagate cache disabled: {log}"
-    );
-    assert!(
-        !log.contains("zccache start")
-            && !log.contains("zccache session-start")
-            && !log.contains("zccache wrapper")
-            && !log.contains("zccache session-end"),
-        "managed zccache should be skipped for non-build cargo commands: {log}"
-    );
-}
-
-#[test]
-fn cargo_front_door_detects_build_after_global_cargo_options() {
-    let cache_root = unique_temp_dir("cargo-global-options-cache");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "--manifest-path", "demo/Cargo.toml", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with global cargo options");
-
-    assert!(
-        output.status.success(),
-        "global-option cargo front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cache=1") && log.contains("zccache start"),
-        "build after global cargo options should still use managed zccache: {log}"
-    );
-}
-
-#[cfg(not(windows))]
-#[test]
-fn cargo_front_door_preserves_jobserver_fds_into_managed_zccache_wrapper() {
-    let cache_root = unique_temp_dir("cargo-jobserver-fds");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_jobserver_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "test", "--no-run"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo test --no-run with fake jobserver fds");
-
-    assert!(
-        output.status.success(),
-        "cache-enabled front door lost jobserver fds\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("failed to connect to jobserver"),
-        "jobserver warning should not be emitted: {stderr}"
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("zccache jobserver fds ok read=3 write=4"),
-        "managed zccache wrapper did not observe open jobserver fds: {log}"
-    );
-}
-
-#[test]
-fn cache_enabled_zccache_build_completes_under_20_seconds() {
-    let cache_root = unique_temp_dir("cargo-zccache-timing");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-
-    let started = Instant::now();
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run soldr cargo build with fake zccache");
-    let elapsed = started.elapsed();
-
-    assert!(
-        output.status.success(),
-        "cache-enabled zccache build failed in {elapsed:?}\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        elapsed < Duration::from_secs(20),
-        "cache-enabled zccache build took {elapsed:?}, expected under 20s"
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("zccache start")
-            && log.contains("zccache session-start")
-            && log.contains("zccache wrapper")
-            && log.contains("zccache session-end test-session"),
-        "timed build should exercise the managed zccache path: {log}"
-    );
-}
-
-#[test]
-fn managed_zccache_rejects_conflicting_cache_dir_override() {
-    let cache_root = unique_temp_dir("cargo-conflicting-zccache-dir");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("ZCCACHE_CACHE_DIR", cache_root.join("user-zccache"))
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .output()
-        .expect("failed to run soldr cargo build with conflicting ZCCACHE_CACHE_DIR");
-
-    assert!(
-        !output.status.success(),
-        "conflicting ZCCACHE_CACHE_DIR should fail"
-    );
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("ZCCACHE_CACHE_DIR is managed by soldr"),
-        "expected explicit override guidance: {stderr}"
-    );
-    assert!(
-        !log_path.exists(),
-        "zccache should not start after a conflicting cache-dir override"
-    );
-}
-
-#[test]
-fn nested_soldr_ignores_inherited_managed_zccache_cache_dir() {
-    let parent_cache_root = unique_temp_dir("cargo-parent-managed-zccache-dir");
-    let child_cache_root = unique_temp_dir("cargo-child-managed-zccache-dir");
-    let parent_zccache_dir = parent_cache_root.join("cache").join("zccache");
-    let child_zccache_dir = child_cache_root.join("cache").join("zccache");
-    let log_path = child_cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &child_cache_root)
-        .env("ZCCACHE_CACHE_DIR", &parent_zccache_dir)
-        .env("SOLDR_MANAGED_ZCCACHE_CACHE_DIR", &parent_zccache_dir)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env_remove("SOLDR_TARGET_CACHE_MODE")
-        .env_remove("SOLDR_BUILD_CACHE_MODE")
-        .output()
-        .expect("failed to run nested soldr cargo build with inherited managed ZCCACHE_CACHE_DIR");
-
-    assert!(
-        output.status.success(),
-        "inherited soldr-managed ZCCACHE_CACHE_DIR should not block nested soldr\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        path_display_variants(&child_zccache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("zccache_dir={path}"))
-                && log.contains(&format!("cache_dir={path}"))),
-        "nested soldr should replace the inherited managed zccache dir with its own cache root: {log}"
-    );
-    assert!(
-        !path_display_variants(&parent_zccache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("cache_dir={path}"))),
-        "nested soldr should not reuse the parent managed zccache dir: {log}"
-    );
-}
-
-#[test]
-fn cargo_front_door_uses_custom_rustc_wrapper_from_env_var() {
-    let cache_root = unique_temp_dir("cargo-custom-wrapper");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let wrapper = install_fake_wrapper(&log_path, "sccache");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_RUSTC_WRAPPER", &wrapper)
-        .output()
-        .expect("failed to run soldr cargo build with custom rustc wrapper");
-
-    assert!(
-        output.status.success(),
-        "custom-wrapper front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains(&format!("cargo wrapper={}", wrapper.display())),
-        "cargo should receive the custom wrapper path: {log}"
-    );
-    assert!(
-        log.contains("sccache wrapper"),
-        "custom wrapper should be invoked for rustc: {log}"
-    );
-    let expected_sccache_dir = cache_root.join("cache").join("sccache");
-    assert!(
-        path_display_variants(&expected_sccache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("sccache_dir={path}"))),
-        "cargo should receive soldr-owned SCCACHE_DIR at {}: {log}",
-        expected_sccache_dir.display()
-    );
-    assert!(
-        expected_sccache_dir.is_dir(),
-        "soldr should pre-create the owned sccache cache dir at {}",
-        expected_sccache_dir.display()
-    );
-    assert!(
-        !log.contains(env!("CARGO_BIN_EXE_soldr")),
-        "soldr should not stay in the wrapper slot when overridden: {log}"
-    );
-    assert!(
-        !log.contains("zccache start")
-            && !log.contains("zccache session-start")
-            && !log.contains("zccache wrapper")
-            && !log.contains("zccache session-end"),
-        "managed zccache should be skipped when using a custom wrapper: {log}"
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        !stderr.contains("soldr: zccache session summary"),
-        "custom wrapper path should not emit zccache session output: {stderr}"
-    );
-}
-
-#[test]
-fn custom_sccache_wrapper_preserves_caller_sccache_dir() {
-    let cache_root = unique_temp_dir("cargo-custom-wrapper-preserve-sccache-dir");
-    let caller_sccache_dir = unique_temp_dir("caller-sccache-dir");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let wrapper = install_fake_wrapper(&log_path, "sccache");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_RUSTC_WRAPPER", &wrapper)
-        .env("SCCACHE_DIR", &caller_sccache_dir)
-        .output()
-        .expect("failed to run soldr cargo build with caller SCCACHE_DIR");
-
-    assert!(
-        output.status.success(),
-        "custom-wrapper front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        path_display_variants(&caller_sccache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("sccache_dir={path}"))),
-        "cargo should preserve caller-provided SCCACHE_DIR at {}: {log}",
-        caller_sccache_dir.display()
-    );
-    let soldr_sccache_dir = cache_root.join("cache").join("sccache");
-    assert!(
-        !path_display_variants(&soldr_sccache_dir)
-            .iter()
-            .any(|path| log.contains(&format!("sccache_dir={path}"))),
-        "cargo should not override caller SCCACHE_DIR with {}: {log}",
-        soldr_sccache_dir.display()
-    );
-}
-
-#[test]
-fn empty_rustc_wrapper_override_disables_wrapper_injection() {
-    let cache_root = unique_temp_dir("cargo-wrapper-disabled");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .env("SOLDR_RUSTC_WRAPPER", "")
-        .output()
-        .expect("failed to run soldr cargo build with wrapper disabled");
-
-    assert!(
-        output.status.success(),
-        "wrapper-disabled front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cargo wrapper= rustc="),
-        "cargo should not receive a wrapper when override is empty: {log}"
-    );
-    assert!(
-        !log.contains("zccache start")
-            && !log.contains("zccache session-start")
-            && !log.contains("zccache wrapper")
-            && !log.contains("zccache session-end"),
-        "managed zccache should be skipped when wrapper injection is disabled: {log}"
-    );
-    assert!(
-        log.contains("rustc ") && log.contains("--crate-name demo"),
-        "rustc should still run directly when wrapper injection is disabled: {log}"
-    );
-}
-
-#[test]
-fn no_cache_bypasses_wrapper_and_zccache() {
-    let cache_root = unique_temp_dir("cargo-no-cache-fake");
-    let log_path = cache_root.join("tool.log");
-    let (cargo, rustc, zccache) = install_fake_toolchain(&log_path);
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "build"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_CARGO_BIN", &cargo)
-        .env("SOLDR_TEST_RUSTC_BIN", &rustc)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .output()
-        .expect("failed to run soldr --no-cache cargo build with fake tools");
-
-    assert!(
-        output.status.success(),
-        "no-cache front door failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("cache=0"),
-        "no-cache front door should propagate cache disable flag: {log}"
-    );
-    assert!(
-        !log.contains("zccache start"),
-        "no-cache front door should not start zccache: {log}"
-    );
-    assert!(
-        !log.contains(env!("CARGO_BIN_EXE_soldr")),
-        "no-cache front door should not set soldr as wrapper: {log}"
-    );
-    assert!(
-        log.contains("rustc ") && log.contains("--crate-name demo"),
-        "no-cache front door should call rustc directly: {log}"
-    );
-}
-
-#[test]
-fn rustc_wrapper_mode_passes_through_to_rustc() {
-    let rustc = rustup_which("rustc");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg(rustc)
-        .arg("--version")
-        .output()
-        .expect("failed to run soldr in wrapper mode");
-
-    assert!(output.status.success(), "wrapper mode failed");
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("rustc"),
-        "unexpected rustc output: {stdout}"
-    );
-}
-
-#[test]
-fn repo_local_toolchain_homes_are_used_when_env_vars_are_unset() {
-    let cache_root = unique_temp_dir("repo-local-toolchain-homes");
-    let log_path = cache_root.join("tool.log");
-    let (rustup, _, _, _) = install_fake_rustup_toolchain(&log_path);
-    let repo_root = unique_temp_dir("repo-local-toolchain-root");
-    let repo_cargo_home = repo_root.join(".cargo");
-    let repo_rustup_home = repo_root.join(".rustup");
-    let nested = repo_root.join("workspace").join("crate");
-    fs::create_dir_all(&repo_cargo_home).expect("failed to create repo-local .cargo");
-    fs::create_dir_all(&repo_rustup_home).expect("failed to create repo-local .rustup");
-    fs::create_dir_all(&nested).expect("failed to create nested working dir");
-
-    for args in [
-        vec!["--no-cache", "cargo", "--version"],
-        vec!["rustfmt", "--version"],
-        vec!["--no-cache", "rustc", "--version"],
-    ] {
-        let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-            .args(&args)
-            .current_dir(&nested)
-            .env("SOLDR_CACHE_DIR", &cache_root)
-            .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-            .env("PATH", isolated_test_path())
-            .env_remove("CARGO_HOME")
-            .env_remove("RUSTUP_HOME")
-            .env_remove("RUSTUP_TOOLCHAIN")
-            .output()
-            .unwrap_or_else(|_| panic!("failed to run soldr with args {args:?}"));
-
-        assert!(
-            output.status.success(),
-            "soldr invocation failed for {:?}\nstdout:\n{}\nstderr:\n{}",
-            args,
-            String::from_utf8_lossy(&output.stdout),
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake rustup log");
-    assert!(
-        log_contains_toolchain_homes(
-            &log,
-            "rustup which cargo",
-            &repo_cargo_home,
-            &repo_rustup_home
-        ),
-        "cargo resolution should use repo-local homes: {log}"
-    );
-    assert!(
-        log_contains_toolchain_homes(&log, "cargo", &repo_cargo_home, &repo_rustup_home),
-        "cargo execution should inherit repo-local homes: {log}"
-    );
-    assert!(
-        log_contains_toolchain_homes(
-            &log,
-            "rustup which rustfmt",
-            &repo_cargo_home,
-            &repo_rustup_home
-        ),
-        "rustfmt resolution should use repo-local homes: {log}"
-    );
-    assert!(
-        log_contains_toolchain_homes(&log, "rustfmt", &repo_cargo_home, &repo_rustup_home),
-        "rustfmt execution should inherit repo-local homes: {log}"
-    );
-    assert!(
-        log_contains_toolchain_homes(
-            &log,
-            "rustup which rustc",
-            &repo_cargo_home,
-            &repo_rustup_home
-        ),
-        "rustc resolution should use repo-local homes: {log}"
-    );
-    assert!(
-        log_contains_toolchain_homes(&log, "rustc", &repo_cargo_home, &repo_rustup_home),
-        "rustc execution should inherit repo-local homes: {log}"
-    );
-}
-
-#[test]
-fn repo_local_cargo_bin_tools_work_without_rustup() {
-    let cache_root = unique_temp_dir("repo-local-cargo-bin");
-    let log_path = cache_root.join("tool.log");
-    let rustup = install_failing_fake_rustup(&log_path);
-    let repo_root = unique_temp_dir("repo-local-cargo-bin-root");
-    let repo_cargo_bin = repo_root.join(".cargo").join("bin");
-    let repo_rustup_home = repo_root.join(".rustup");
-    let nested = repo_root.join("workspace").join("crate");
-    fs::create_dir_all(&repo_cargo_bin).expect("failed to create repo-local .cargo/bin");
-    // Anchor the rustup-home ancestor walk inside the test sandbox so it can't
-    // climb up to a runner-installed `~/.rustup` (Windows GitHub runners put
-    // TEMP under USERPROFILE, where `.rustup` typically exists).
-    fs::create_dir_all(&repo_rustup_home).expect("failed to create repo-local .rustup");
-    fs::create_dir_all(&nested).expect("failed to create nested working dir");
-    install_fake_version_toolchain(&repo_cargo_bin, &log_path);
-
-    for args in [
-        vec!["--no-cache", "cargo", "--version"],
-        vec!["rustfmt", "--version"],
-        vec!["--no-cache", "rustc", "--version"],
-    ] {
-        let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-            .args(&args)
-            .current_dir(&nested)
-            .env("SOLDR_CACHE_DIR", &cache_root)
-            .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-            .env("PATH", isolated_test_path())
-            .env_remove("CARGO_HOME")
-            .env_remove("RUSTUP_HOME")
-            .env_remove("RUSTUP_TOOLCHAIN")
-            .output()
-            .unwrap_or_else(|_| panic!("failed to run soldr with args {args:?}"));
-
-        assert!(
-            output.status.success(),
-            "soldr invocation failed for {:?}\nstdout:\n{}\nstderr:\n{}",
-            args,
-            String::from_utf8_lossy(&output.stdout),
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.lines().any(|line| line.starts_with("cargo ")),
-        "expected repo-local cargo shim to run: {log}"
-    );
-    assert!(
-        log.lines().any(|line| line.starts_with("rustfmt ")),
-        "expected repo-local rustfmt shim to run: {log}"
-    );
-    assert!(
-        log.lines().any(|line| line.starts_with("rustc ")),
-        "expected repo-local rustc shim to run: {log}"
-    );
-    assert!(
-        !log.lines().any(|line| line.starts_with("rustup ")),
-        "repo-local .cargo/bin tools should bypass rustup entirely: {log}"
-    );
-}
-
-#[test]
-fn explicit_toolchain_home_env_vars_win_over_repo_local_homes() {
-    let cache_root = unique_temp_dir("explicit-toolchain-homes");
-    let log_path = cache_root.join("tool.log");
-    let (rustup, _, _, _) = install_fake_rustup_toolchain(&log_path);
-    let repo_root = unique_temp_dir("explicit-toolchain-repo");
-    let repo_cargo_home = repo_root.join(".cargo");
-    let repo_rustup_home = repo_root.join(".rustup");
-    let nested = repo_root.join("workspace").join("crate");
-    let explicit_cargo_home = unique_temp_dir("explicit-cargo-home");
-    let explicit_rustup_home = unique_temp_dir("explicit-rustup-home");
-    fs::create_dir_all(&repo_cargo_home).expect("failed to create repo-local .cargo");
-    fs::create_dir_all(&repo_rustup_home).expect("failed to create repo-local .rustup");
-    fs::create_dir_all(&nested).expect("failed to create nested working dir");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "--version"])
-        .current_dir(&nested)
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .env("CARGO_HOME", &explicit_cargo_home)
-        .env("RUSTUP_HOME", &explicit_rustup_home)
-        .env("PATH", isolated_test_path())
-        .env_remove("RUSTUP_TOOLCHAIN")
-        .output()
-        .expect("failed to run soldr cargo --version with explicit homes");
-
-    assert!(
-        output.status.success(),
-        "soldr cargo --version failed with explicit homes\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake rustup log");
-    let explicit_cargo_home = explicit_cargo_home.display().to_string();
-    let explicit_rustup_home = explicit_rustup_home.display().to_string();
-    assert!(
-        log.contains(&format!(
-            "rustup which cargo cargo_home={explicit_cargo_home} rustup_home={explicit_rustup_home}"
-        )),
-        "cargo resolution should prefer explicit homes: {log}"
-    );
-    assert!(
-        log.contains(&format!(
-            "cargo cargo_home={explicit_cargo_home} rustup_home={explicit_rustup_home}"
-        )),
-        "cargo execution should inherit explicit homes: {log}"
-    );
-    assert!(
-        !log.contains(&repo_cargo_home.display().to_string())
-            && !log.contains(&repo_rustup_home.display().to_string()),
-        "repo-local homes should not leak into explicit-home runs: {log}"
-    );
-}
-
-#[test]
-fn rustup_resolution_failure_reports_raw_error_and_ci_guidance() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "rustc", "--version"])
-        .env("RUSTUP_TOOLCHAIN", "soldr-ci-missing-toolchain")
-        .output()
-        .expect("failed to run soldr --no-cache rustc --version with invalid RUSTUP_TOOLCHAIN");
-
-    assert!(
-        !output.status.success(),
-        "expected soldr rustc --version to fail when rustup resolution fails"
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("failed to resolve rustc via rustup: error: override toolchain 'soldr-ci-missing-toolchain' is not installed"),
-        "expected raw rustup stderr to be preserved: {stderr}"
-    );
-    assert!(
-        stderr.contains(
-            "the RUSTUP_TOOLCHAIN environment variable specifies an uninstalled toolchain"
-        ),
-        "expected raw rustup explanation in stderr: {stderr}"
-    );
-    assert!(
-        stderr.contains("pins Rust in rust-toolchain.toml"),
-        "expected rust-toolchain.toml guidance in stderr: {stderr}"
-    );
-    assert!(
-        stderr.contains("generic stable toolchain"),
-        "expected exact-channel guidance in stderr: {stderr}"
-    );
-    assert!(
-        stderr.contains("RUSTUP_TOOLCHAIN"),
-        "expected RUSTUP_TOOLCHAIN guidance in stderr: {stderr}"
-    );
-    assert!(
-        stderr.contains("setup-soldr action path"),
-        "expected setup-soldr guidance in stderr: {stderr}"
-    );
-}
-
-#[test]
-fn status_reports_cache_control_defaults() {
-    let cache_root = unique_temp_dir("status");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("status")
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr status");
-
-    assert!(output.status.success(), "status command failed");
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("cache dir:"),
-        "status missing cache dir: {stdout}"
-    );
-    assert!(
-        stdout.contains("cache default: enabled"),
-        "status missing cache default: {stdout}"
-    );
-    assert!(
-        stdout.contains("zccache version:"),
-        "status missing zccache version: {stdout}"
-    );
-    assert!(
-        stdout.contains("soldr zccache cache dir:"),
-        "status missing effective zccache cache dir: {stdout}"
-    );
-    assert!(
-        stdout.contains("not fetched yet"),
-        "status should explain unfetched managed zccache state: {stdout}"
-    );
-}
-
-#[test]
-fn status_json_reports_stable_machine_fields() {
-    let cache_root = unique_temp_dir("status-json");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["status", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr status --json");
-
-    assert!(output.status.success(), "status --json command failed");
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("status --json did not return JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "status");
-    assert_eq!(json["soldr_version"], env!("CARGO_PKG_VERSION"));
-    assert_eq!(json["cache_default_enabled"], true);
-    assert_eq!(json["cache_enabled_for_invocation"], true);
-    assert_eq!(json["managed_zccache_version"], "1.5.0");
-    assert_eq!(json["root_dir"], cache_root.display().to_string());
-    assert_eq!(
-        json["cache_dir"],
-        cache_root.join("cache").display().to_string()
-    );
-    assert_eq!(json["zccache"]["binary_fetched"], false);
-    assert_eq!(json["zccache"]["session_log_present"], false);
-    assert_eq!(json["zccache"]["journal_present"], false);
-    assert_eq!(json["zccache"]["session_stats_present"], false);
-    assert_eq!(
-        json["zccache"]["cache_dir"],
-        cache_root
-            .join("cache")
-            .join("zccache")
-            .display()
-            .to_string()
-    );
-    assert!(
-        json["target"].as_str().is_some(),
-        "status JSON missing target"
-    );
-    assert_eq!(
-        json["zccache"]["session_log_path"],
-        cache_root
-            .join("cache")
-            .join("zccache")
-            .join("logs")
-            .join("last-session.log")
-            .display()
-            .to_string()
-    );
-    assert_eq!(
-        json["zccache"]["session_stats_path"],
-        cache_root
-            .join("cache")
-            .join("zccache")
-            .join("logs")
-            .join("last-session-stats.json")
-            .display()
-            .to_string()
-    );
-}
-
-#[test]
-fn cache_command_reports_managed_zccache_status() {
-    let cache_root = unique_temp_dir("cache-command");
-    let log_path = cache_root.join("tool.log");
-    let (_, _, zccache) = install_fake_toolchain(&log_path);
-    let journal = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("logs")
-        .join("last-session.jsonl");
-    let session_log = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("logs")
-        .join("last-session.log");
-    let session_stats = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("logs")
-        .join("last-session-stats.json");
-    fs::create_dir_all(journal.parent().expect("journal parent missing"))
-        .expect("failed to create journal dir");
-    fs::write(&session_log, "compile line\n").expect("failed to seed session log");
-    fs::write(&journal, "{\"event\":\"hit\"}\n").expect("failed to seed journal");
-    fs::write(
-        &session_stats,
-        "{\"status\":\"ok\",\"hits\":7,\"misses\":3}\n",
-    )
-    .expect("failed to seed session stats");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("cache")
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .output()
-        .expect("failed to run soldr cache");
-
-    assert!(output.status.success(), "cache command failed");
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("soldr zccache cache dir:"),
-        "cache command missing cache dir: {stdout}"
-    );
-    assert!(
-        stdout.contains("soldr zccache state dir:"),
-        "cache command missing state dir: {stdout}"
-    );
-    assert!(
-        stdout.contains("last session log:"),
-        "cache command missing session log path: {stdout}"
-    );
-    assert!(
-        stdout.contains("last session journal:"),
-        "cache command missing journal path: {stdout}"
-    );
-    assert!(
-        stdout.contains("last session stats:"),
-        "cache command missing session stats path: {stdout}"
-    );
-    assert!(
-        stdout.contains(&format!("{}", session_log.display())) && stdout.contains("(present)"),
-        "cache command should report present session log: {stdout}"
-    );
-    assert!(
-        stdout.contains("zccache: hits=7"),
-        "cache command should surface managed zccache status output: {stdout}"
-    );
-}
-
-#[test]
-fn cache_json_reports_managed_zccache_status() {
-    let cache_root = unique_temp_dir("cache-command-json");
-    let log_path = cache_root.join("tool.log");
-    let (_, _, zccache) = install_fake_toolchain(&log_path);
-    let journal = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("logs")
-        .join("last-session.jsonl");
-    let session_log = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("logs")
-        .join("last-session.log");
-    let session_stats = cache_root
-        .join("cache")
-        .join("zccache")
-        .join("logs")
-        .join("last-session-stats.json");
-    fs::create_dir_all(journal.parent().expect("journal parent missing"))
-        .expect("failed to create journal dir");
-    fs::write(&session_log, "compile line\n").expect("failed to seed session log");
-    fs::write(&journal, "{\"event\":\"hit\"}\n").expect("failed to seed journal");
-    fs::write(
-        &session_stats,
-        "{\"status\":\"ok\",\"hits\":7,\"misses\":3}\n",
-    )
-    .expect("failed to seed session stats");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cache", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .output()
-        .expect("failed to run soldr cache --json");
-
-    assert!(output.status.success(), "cache --json command failed");
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("cache --json did not return JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "cache");
-    assert_eq!(json["managed_zccache_version"], "1.5.0");
-    assert_eq!(json["zccache"]["session_log_present"], true);
-    assert_eq!(json["zccache"]["journal_present"], true);
-    assert_eq!(json["zccache"]["session_stats_present"], true);
-    assert_eq!(json["zccache"]["binary_fetched"], true);
-    assert_eq!(
-        json["zccache"]["cache_dir"],
-        cache_root
-            .join("cache")
-            .join("zccache")
-            .display()
-            .to_string()
-    );
-    assert_eq!(
-        json["zccache"]["session_log_path"],
-        session_log.display().to_string()
-    );
-    assert_eq!(
-        json["zccache"]["journal_path"],
-        journal.display().to_string()
-    );
-    assert_eq!(
-        json["zccache"]["session_stats_path"],
-        session_stats.display().to_string()
-    );
-    assert_eq!(
-        json["zccache"]["status_lines"][0],
-        Value::String("hits=7".to_string())
-    );
-}
-
-#[test]
-fn cache_report_json_emits_stable_schema_when_files_missing() {
-    let cache_root = unique_temp_dir("cache-report-empty");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cache", "report", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr cache report --json");
-
-    assert!(
-        output.status.success(),
-        "cache report --json failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value = serde_json::from_slice(&output.stdout)
-        .expect("cache report --json must produce parseable JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "cache report");
-    assert_eq!(json["managed_zccache_version"], "1.5.0");
-    assert_eq!(json["session_stats_present"], false);
-    assert_eq!(json["journal_present"], false);
-    assert!(json["last_session"].is_null());
-    assert!(json["rollups"].is_null());
-    assert!(
-        json["diagnoses"]
-            .as_array()
-            .expect("diagnoses array")
-            .is_empty(),
-        "diagnoses should start empty before rule passes are wired"
-    );
-    let notes = json["notes"]
-        .as_array()
-        .expect("notes should always be an array");
-    assert!(
-        !notes.is_empty(),
-        "missing files should produce explanatory notes"
-    );
-}
-
-#[test]
-fn cache_report_json_surfaces_persisted_session_stats() {
-    let cache_root = unique_temp_dir("cache-report-stats");
-    let zccache_dir = cache_root.join("cache").join("zccache");
-    let stats_path = zccache_dir.join("logs").join("last-session-stats.json");
-    fs::create_dir_all(stats_path.parent().expect("stats parent missing"))
-        .expect("failed to create logs dir");
-    fs::write(
-        &stats_path,
-        r#"{"status":"ok","session_id":"abc","hits":7,"misses":3,"compilations":10,"hit_rate":0.7,"time_saved_ms":12345}"#,
-    )
-    .expect("failed to seed session stats");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cache", "report", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr cache report --json");
-
-    assert!(
-        output.status.success(),
-        "cache report --json failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("cache report --json must produce JSON");
-    assert_eq!(json["session_stats_present"], true);
-    assert_eq!(json["last_session"]["status"], "ok");
-    assert_eq!(json["last_session"]["hits"], 7);
-    assert_eq!(json["last_session"]["misses"], 3);
-    assert_eq!(json["last_session"]["hit_rate"].as_f64(), Some(0.7));
-}
-
-#[test]
-fn clean_clears_managed_zccache_and_state_dir() {
-    let cache_root = unique_temp_dir("clean-command");
-    let log_path = cache_root.join("tool.log");
-    let (_, _, zccache) = install_fake_toolchain(&log_path);
-    let state_dir = cache_root.join("cache").join("zccache");
-    let user_home = cache_root.join("user-home");
-    let user_global_zccache = user_home.join(".zccache");
-    fs::create_dir_all(&user_global_zccache).expect("failed to seed user-global zccache");
-    fs::write(user_global_zccache.join("index.redb"), "user cache")
-        .expect("failed to seed user-global zccache file");
-    let journal = state_dir.join("logs").join("last-session.jsonl");
-    fs::create_dir_all(journal.parent().expect("journal parent missing"))
-        .expect("failed to create journal dir");
-    fs::write(&journal, "{\"event\":\"hit\"}\n").expect("failed to seed journal");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("clean")
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("HOME", &user_home)
-        .env("USERPROFILE", &user_home)
-        .env("SOLDR_TEST_ZCCACHE_BIN", &zccache)
-        .output()
-        .expect("failed to run soldr clean");
-
-    assert!(output.status.success(), "clean command failed");
-    assert!(
-        !state_dir.exists(),
-        "clean should remove soldr zccache state dir at {}",
-        state_dir.display()
-    );
-    assert!(
-        user_global_zccache.join("index.redb").exists(),
-        "clean must not remove user-global zccache state at {}",
-        user_global_zccache.display()
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("cleared zccache artifact cache"),
-        "clean should report artifact cache cleanup: {stdout}"
-    );
-    assert!(
-        stdout.contains("removed soldr zccache state dir:"),
-        "clean should report state dir cleanup: {stdout}"
-    );
-
-    let log = fs::read_to_string(&log_path).expect("failed to read fake tool log");
-    assert!(
-        log.contains("zccache clear"),
-        "clean should call managed zccache clear: {log}"
-    );
-}
-
-#[test]
-fn purge_removes_soldr_artifact_dirs_and_keeps_config() {
-    let cache_root = unique_temp_dir("purge-command");
-    let bin_dir = cache_root.join("bin");
-    let cache_dir = cache_root.join("cache");
-    let zccache_state_dir = cache_dir.join("zccache").join("logs");
-    let config_file = cache_root.join("config.toml");
-    fs::create_dir_all(&bin_dir).expect("failed to create bin dir");
-    fs::create_dir_all(&zccache_state_dir).expect("failed to create zccache state dir");
-    fs::write(bin_dir.join("soldr-tool"), "cached binary").expect("failed to seed bin cache");
-    fs::write(zccache_state_dir.join("last-session.jsonl"), "{}\n")
-        .expect("failed to seed zccache state");
-    fs::write(&config_file, "cache = true\n").expect("failed to seed config");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("purge")
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr purge");
-
-    assert!(output.status.success(), "purge command failed");
-    assert!(
-        !bin_dir.exists(),
-        "purge should remove soldr-managed fetched tool artifacts at {}",
-        bin_dir.display()
-    );
-    assert!(
-        !cache_dir.exists(),
-        "purge should remove soldr-managed cache artifacts at {}",
-        cache_dir.display()
-    );
-    assert!(
-        config_file.exists(),
-        "purge should keep non-artifact config at {}",
-        config_file.display()
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("removed soldr cache dir:"),
-        "purge should report cache cleanup: {stdout}"
-    );
-    assert!(
-        stdout.contains("removed soldr bin dir:"),
-        "purge should report bin cleanup: {stdout}"
-    );
-}
-
-#[test]
-fn purge_reports_empty_cache_without_creating_dirs() {
-    let cache_root = unique_temp_dir("purge-empty-command");
-    let bin_dir = cache_root.join("bin");
-    let cache_dir = cache_root.join("cache");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("purge")
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr purge");
-
-    assert!(output.status.success(), "purge command failed");
-    assert!(
-        !bin_dir.exists(),
-        "purge should not create missing bin dir at {}",
-        bin_dir.display()
-    );
-    assert!(
-        !cache_dir.exists(),
-        "purge should not create missing cache dir at {}",
-        cache_dir.display()
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("soldr cache is already empty:"),
-        "purge should report empty cache: {stdout}"
-    );
-}
-
-#[test]
-fn purge_removes_corrupt_artifact_paths() {
-    let cache_root = unique_temp_dir("purge-corrupt-command");
-    let bin_path = cache_root.join("bin");
-    let cache_path = cache_root.join("cache");
-    fs::write(&bin_path, "not a dir").expect("failed to seed corrupt bin path");
-    fs::write(&cache_path, "not a dir").expect("failed to seed corrupt cache path");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .arg("purge")
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .output()
-        .expect("failed to run soldr purge");
-
-    assert!(output.status.success(), "purge command failed");
-    assert!(
-        !bin_path.exists(),
-        "purge should remove corrupt soldr bin path at {}",
-        bin_path.display()
-    );
-    assert!(
-        !cache_path.exists(),
-        "purge should remove corrupt soldr cache path at {}",
-        cache_path.display()
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("removed soldr cache entry:"),
-        "purge should report corrupt cache path cleanup: {stdout}"
-    );
-    assert!(
-        stdout.contains("removed soldr bin entry:"),
-        "purge should report corrupt bin path cleanup: {stdout}"
-    );
-}
-
-#[test]
-fn purge_rejects_json_flag() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["purge", "--json"])
-        .output()
-        .expect("failed to run soldr purge --json");
-
-    assert!(
-        !output.status.success(),
-        "purge --json should be rejected because JSON is not supported there"
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("--json"),
-        "expected clap to reject purge --json: {stderr}"
-    );
-}
-
-#[test]
-fn clean_rejects_json_flag() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["clean", "--json"])
-        .output()
-        .expect("failed to run soldr clean --json");
-
-    assert!(
-        !output.status.success(),
-        "clean --json should be rejected because JSON is not supported there"
-    );
-
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("--json"),
-        "expected clap to reject clean --json: {stderr}"
-    );
-}
-
-#[cfg(windows)]
-#[test]
-fn cargo_front_door_forces_msvc_target_even_with_polluted_path() {
-    let fake_tools = unique_temp_dir("fake-tools");
-    fs::write(
-        fake_tools.join("cargo.cmd"),
-        "@echo off\r\necho fake cargo should not be used 1>&2\r\nexit /b 1\r\n",
-    )
-    .expect("failed to write fake cargo.cmd");
-    fs::write(
-        fake_tools.join("rustc.cmd"),
-        "@echo off\r\necho fake rustc should not be used 1>&2\r\nexit /b 1\r\n",
-    )
-    .expect("failed to write fake rustc.cmd");
-
-    let target_dir = unique_temp_dir("target-dir");
-    let fixture = PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .join("tests")
-        .join("fixtures")
-        .join("windows-msvc-default");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["--no-cache", "cargo", "build"])
-        .current_dir(&fixture)
-        .env("PATH", prepend_to_path(&fake_tools))
-        .env("CARGO_TARGET_DIR", &target_dir)
-        .env("SOLDR_CACHE_DIR", unique_temp_dir("msvc-cache-root"))
-        .output()
-        .expect("failed to run soldr cargo build");
-
-    assert!(
-        output.status.success(),
-        "soldr cargo build failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let artifact = target_dir
-        .join("x86_64-pc-windows-msvc")
-        .join("debug")
-        .join("windows-msvc-default.exe");
-    assert!(
-        artifact.exists(),
-        "expected MSVC target artifact at {}",
-        artifact.display()
-    );
-}
-
-/// Regression test for issue #324: RUSTC_WRAPPER mode must propagate the
-/// rustc exit code even when the source is read from stdin ("-").
-///
-/// Before the fix, zccache consumed stdin to hash the source. Rustc then
-/// received an empty stdin, compiled nothing, and exited 0 — masking E0554
-/// and similar errors that build-script feature probes depend on.
-///
-/// The test exercises the direct-to-rustc path (no zccache configured) to
-/// keep it self-contained and fast. The spill-to-tempfile logic runs
-/// regardless of whether zccache is in the chain.
-#[test]
-fn wrapper_mode_stdin_source_propagates_nonzero_exit_code() {
-    let rustc = rustup_which("rustc");
-    let out_dir = unique_temp_dir("wrapper-stdin-exit");
-
-    // A source that is valid Rust but uses an unstable feature gate.
-    // On stable rustc this must fail with E0554 (exit != 0).
-    let probe_source = b"#![allow(stable_features)]\n#![feature(rustc_attrs)]\n";
-
-    // Invoke soldr as RUSTC_WRAPPER: soldr <rustc-path> - <flags...>
-    // Disable the cache to avoid a zccache binary being required.
-    let mut child = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args([
-            rustc.as_str(),
-            "-",
-            "--crate-type=lib",
-            "--emit=metadata",
-            "--out-dir",
-            out_dir.to_str().expect("non-UTF-8 temp dir"),
-        ])
-        .env("SOLDR_CACHE_ENABLED", "0")
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()
-        .expect("failed to spawn soldr in wrapper mode");
-
-    child
-        .stdin
-        .take()
-        .expect("stdin should be piped")
-        .write_all(probe_source)
-        .expect("failed to write probe source to soldr stdin");
-
-    let output = child
-        .wait_with_output()
-        .expect("failed to wait for soldr wrapper");
-
-    assert!(
-        !output.status.success(),
-        "soldr wrapper mode with stdin source must propagate non-zero rustc exit code\n\
-         stdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr),
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Issue #323 — soldr gc cargo / gc locations / gc sweep coverage.
-// ---------------------------------------------------------------------------
-
-#[test]
-fn gc_cargo_help_lists_max_flags_and_dry_run() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "cargo", "--help"])
-        .output()
-        .expect("failed to run soldr gc cargo --help");
-    assert!(output.status.success(), "soldr gc cargo --help failed");
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    for flag in [
-        "--dry-run",
-        "--toolchain",
-        "--max-src-age",
-        "--max-crate-age",
-        "--max-index-age",
-        "--max-git-co-age",
-        "--max-git-db-age",
-        "--max-download-age",
-        "--max-src-size",
-        "--max-crate-size",
-        "--max-git-size",
-        "--max-download-size",
-        "--json",
-    ] {
-        assert!(
-            stdout.contains(flag),
-            "gc cargo --help missing {flag}: {stdout}"
-        );
-    }
-}
-
-#[test]
-fn gc_cargo_rejects_unknown_flag() {
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "cargo", "--bogus-flag"])
-        .output()
-        .expect("failed to run soldr gc cargo --bogus-flag");
-    assert!(!output.status.success(), "unknown flag should fail");
-    let stderr = String::from_utf8_lossy(&output.stderr);
-    assert!(
-        stderr.contains("unexpected argument") || stderr.contains("unrecognized"),
-        "expected clap-style unknown-flag error, got: {stderr}"
-    );
-}
-
-#[test]
-fn gc_locations_json_emits_valid_schema_even_without_caches() {
-    let cache_root = unique_temp_dir("gc-locations-json");
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "locations", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        // Disable auto-GC so the test isn't affected by background work.
-        .env("SOLDR_AUTO_GC_DISABLED", "1")
-        .output()
-        .expect("failed to run soldr gc locations --json");
-
-    assert!(
-        output.status.success(),
-        "gc locations failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("gc locations --json must be JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "gc");
-    assert_eq!(json["mode"], "locations");
-    let locations = json["locations"]
-        .as_array()
-        .expect("locations must be an array");
-    assert!(!locations.is_empty(), "should report at least soldr_cache");
-    // The kinds we expect to always appear, even when paths don't exist.
-    let kinds: Vec<&str> = locations
-        .iter()
-        .filter_map(|entry| entry["kind"].as_str())
-        .collect();
-    for required in [
-        "cargo_registry_src",
-        "cargo_registry_cache",
-        "cargo_registry_index",
-        "cargo_git_db",
-        "cargo_git_checkouts",
-        "cargo_global_cache",
-        "rustup_toolchains",
-        "rustup_update_hashes",
-        "soldr_cache",
-        "soldr_state_db",
-    ] {
-        assert!(
-            kinds.contains(&required),
-            "gc locations missing kind {required}: {kinds:?}"
-        );
-    }
-    // Each entry must carry the schema fields.
-    for entry in locations {
-        for field in [
-            "kind",
-            "path",
-            "exists",
-            "size_bytes",
-            "size_human",
-            "file_count",
-            "owner",
-            "purge_safety",
-        ] {
-            assert!(
-                entry.get(field).is_some(),
-                "gc locations entry missing field {field}: {entry}"
-            );
-        }
-    }
-}
-
-#[test]
-fn gc_sweep_no_cargo_dry_run_runs_end_to_end_without_changes() {
-    let cache_root = unique_temp_dir("gc-sweep-dryrun");
-    let target = seed_gc_candidate(&cache_root, "sweep-dryrun");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["gc", "sweep", "--no-cargo-gc", "--dry-run", "--json"])
-        .env("SOLDR_CACHE_DIR", &cache_root)
-        .env("SOLDR_AUTO_GC_DISABLED", "1")
-        .output()
-        .expect("failed to run soldr gc sweep --no-cargo-gc --dry-run --json");
-
-    assert!(
-        output.status.success(),
-        "gc sweep dry-run failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        target.exists(),
-        "dry-run sweep must not delete {}",
-        target.display()
-    );
-
-    let json: Value = serde_json::from_slice(&output.stdout).expect("gc sweep --json must be JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "gc");
-    assert_eq!(json["mode"], "sweep");
-    assert_eq!(json["dry_run"], true);
-    assert!(
-        json["cargo_gc"].is_null(),
-        "cargo_gc must be null when --no-cargo-gc"
-    );
-    assert!(
-        json["soldr_targets"].is_null(),
-        "dry-run sweep must not run soldr target purge"
-    );
-    assert!(json["locations"].is_array());
-}
-
-#[test]
-fn cache_prune_target_dry_run_emits_json() {
-    let cache_root = unique_temp_dir("cache-prune-target-json");
-    let target = cache_root.join("target");
-    let parent = target.join("debug").join("deps");
-    fs::create_dir_all(&parent).expect("failed to create deps dir");
-    // Use file entries (with the cargo-style `.rlib` extension) rather
-    // than directories — `set_modified` on a directory handle requires
-    // privileged access on Windows.
-    let older = parent.join("libfoo-aaaaaaaaaaaaa.rlib");
-    let newer = parent.join("libfoo-bbbbbbbbbbbbb.rlib");
-    fs::write(&older, b"older").expect("failed to seed older content");
-    fs::write(&newer, b"newer-with-more-bytes").expect("failed to seed newer content");
-
-    // Pin mtimes so the test is deterministic on slow filesystems.
-    let older_when = SystemTime::UNIX_EPOCH + Duration::from_secs(1_700_000_000);
-    let newer_when = SystemTime::UNIX_EPOCH + Duration::from_secs(1_700_000_500);
-    {
-        let f = fs::File::options()
-            .write(true)
-            .open(&older)
-            .expect("open older for mtime");
-        f.set_modified(older_when).expect("set older mtime");
-    }
-    {
-        let f = fs::File::options()
-            .write(true)
-            .open(&newer)
-            .expect("open newer for mtime");
-        f.set_modified(newer_when).expect("set newer mtime");
-    }
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["cache", "prune-target"])
-        .arg(&target)
-        .args(["--dry-run", "--json"])
-        .output()
-        .expect("failed to run soldr cache prune-target --dry-run --json");
-
-    assert!(
-        output.status.success(),
-        "cache prune-target failed\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    // Dry-run must leave everything on disk.
-    assert!(older.exists(), "dry-run must not delete older entry");
-    assert!(newer.exists(), "dry-run must not delete newer entry");
-
-    let json: Value = serde_json::from_slice(&output.stdout)
-        .expect("cache prune-target --json must produce parseable JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "cache prune-target");
-    assert_eq!(json["dry_run"], true);
-    assert_eq!(json["scanned"], 2);
-    assert_eq!(json["kept"], 1);
-    assert_eq!(json["deleted"], 1);
-
-    let entries = json["entries"]
-        .as_array()
-        .expect("entries must be an array");
-    assert_eq!(entries.len(), 2);
-    let mut found_keep = false;
-    let mut found_delete = false;
-    for entry in entries {
-        for field in [
-            "path",
-            "prefix",
-            "hash",
-            "size_bytes",
-            "size_human",
-            "mtime_unix",
-            "action",
-        ] {
-            assert!(
-                entry.get(field).is_some(),
-                "prune-target entry missing field {field}: {entry}"
-            );
-        }
-        assert_eq!(entry["prefix"], "libfoo");
-        match entry["action"].as_str() {
-            Some("keep") => {
-                found_keep = true;
-                assert_eq!(entry["hash"], "bbbbbbbbbbbbb");
-            }
-            Some("delete") => {
-                found_delete = true;
-                assert_eq!(entry["hash"], "aaaaaaaaaaaaa");
-            }
-            other => panic!("unexpected action {other:?} in {entry}"),
-        }
-    }
-    assert!(found_keep, "must have one keep entry");
-    assert!(found_delete, "must have one delete entry");
-}
-
-/// Canned output for `install_doctor_fake_rustup`. Each field maps to one
-/// of the three rustup subcommands invoked by `soldr doctor`. Lines are
-/// written verbatim into the corresponding response files.
-struct DoctorFakeRustupBehavior {
-    toolchain_list: Vec<String>,
-    components_installed: Vec<String>,
-    targets_installed: Vec<String>,
-}
-
-/// Fake rustup that branches on argv to satisfy `soldr doctor` queries:
-/// - `toolchain list` echoes one line per `behavior.toolchain_list`
-/// - `component list --installed --toolchain <ch>` echoes one line per
-///   `behavior.components_installed`
-/// - `target list --installed --toolchain <ch>` echoes one line per
-///   `behavior.targets_installed`
-///
-/// Any other invocation prints an error and exits 1 (the test must fail).
-/// Each invocation is also logged to `log_path` like the logging fake.
-fn install_doctor_fake_rustup(log_path: &Path, behavior: &DoctorFakeRustupBehavior) -> PathBuf {
-    let dir = unique_temp_dir("fake-rustup-doctor");
-    let toolchain_list_path = dir.join("toolchain_list.txt");
-    let components_path = dir.join("components.txt");
-    let targets_path = dir.join("targets.txt");
-    fs::write(&toolchain_list_path, behavior.toolchain_list.join("\n"))
-        .expect("failed to write fake toolchain list");
-    fs::write(&components_path, behavior.components_installed.join("\n"))
-        .expect("failed to write fake components list");
-    fs::write(&targets_path, behavior.targets_installed.join("\n"))
-        .expect("failed to write fake targets list");
-
-    #[cfg(windows)]
-    let rustup = dir.join("rustup.bat");
-    #[cfg(not(windows))]
-    let rustup = fake_script_path(&dir, "rustup");
-
-    let script = doctor_fake_rustup_script(
-        log_path,
-        &toolchain_list_path,
-        &components_path,
-        &targets_path,
-    );
-    write_fake_script(&rustup, &script);
-    rustup
-}
-
-fn doctor_fake_rustup_script(
-    log_path: &Path,
-    toolchain_list_path: &Path,
-    components_path: &Path,
-    targets_path: &Path,
-) -> String {
-    #[cfg(windows)]
-    {
-        format!(
-            "@echo off\n\
-             setlocal enabledelayedexpansion\n\
-             set \"first=%~1\"\n\
-             set \"second=%~2\"\n\
-             set \"line=\"\n\
-             :loop\n\
-             if \"%~1\"==\"\" goto done\n\
-             if defined line (set \"line=!line!\u{1f}%~1\") else (set \"line=%~1\")\n\
-             shift\n\
-             goto loop\n\
-             :done\n\
-             echo !line!>>\"{log}\"\n\
-             if /I \"!first!\"==\"toolchain\" (\n\
-               if /I \"!second!\"==\"list\" (\n\
-                 type \"{toolchain_list}\"\n\
-                 exit /b 0\n\
-               )\n\
-             )\n\
-             if /I \"!first!\"==\"component\" (\n\
-               if /I \"!second!\"==\"list\" (\n\
-                 type \"{components}\"\n\
-                 exit /b 0\n\
-               )\n\
-             )\n\
-             if /I \"!first!\"==\"target\" (\n\
-               if /I \"!second!\"==\"list\" (\n\
-                 type \"{targets}\"\n\
-                 exit /b 0\n\
-               )\n\
-             )\n\
-             echo unsupported rustup invocation 1>&2\n\
-             exit /b 1\n",
-            log = log_path.display(),
-            toolchain_list = toolchain_list_path.display(),
-            components = components_path.display(),
-            targets = targets_path.display(),
-        )
-    }
-    #[cfg(not(windows))]
-    {
-        format!(
-            "#!/bin/sh\n\
-             sep=$(printf '\\037')\n\
-             out=\"\"\n\
-             first=1\n\
-             for arg in \"$@\"; do\n\
-               if [ $first -eq 1 ]; then\n\
-                 out=\"$arg\"\n\
-                 first=0\n\
-               else\n\
-                 out=\"$out${{sep}}$arg\"\n\
-               fi\n\
-             done\n\
-             printf '%s\\n' \"$out\" >> \"{log}\"\n\
-             if [ \"$1\" = \"toolchain\" ] && [ \"$2\" = \"list\" ]; then\n\
-               cat \"{toolchain_list}\"\n\
-               exit 0\n\
-             fi\n\
-             if [ \"$1\" = \"component\" ] && [ \"$2\" = \"list\" ]; then\n\
-               cat \"{components}\"\n\
-               exit 0\n\
-             fi\n\
-             if [ \"$1\" = \"target\" ] && [ \"$2\" = \"list\" ]; then\n\
-               cat \"{targets}\"\n\
-               exit 0\n\
-             fi\n\
-             echo \"unsupported rustup invocation: $*\" >&2\n\
-             exit 1\n",
-            log = log_path.display(),
-            toolchain_list = toolchain_list_path.display(),
-            components = components_path.display(),
-            targets = targets_path.display(),
-        )
-    }
-}
-
-#[test]
-fn doctor_reports_drift_when_component_missing() {
-    let workspace = unique_temp_dir("doctor-drift-component");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         components = [\"clippy\"]\n",
-    );
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_doctor_fake_rustup(
-        &log_path,
-        &DoctorFakeRustupBehavior {
-            toolchain_list: vec!["1.94.1-x86_64-unknown-linux-gnu (default)".to_string()],
-            components_installed: Vec::new(),
-            targets_installed: vec!["x86_64-unknown-linux-gnu".to_string()],
-        },
-    );
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["doctor", "--json"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr doctor --json");
-
-    assert_eq!(
-        output.status.code(),
-        Some(1),
-        "doctor must exit 1 when drift detected\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("doctor --json must produce JSON");
-    assert_eq!(json["schema_version"], 1);
-    assert_eq!(json["command"], "doctor");
-    assert_eq!(json["toolchain"]["channel"], "1.94.1");
-    assert_eq!(json["toolchain"]["installed"], true);
-    assert_eq!(json["drift"], true);
-    assert_eq!(
-        json["missing_components"],
-        serde_json::json!(["clippy"]),
-        "missing_components must list clippy"
-    );
-    assert_eq!(
-        json["missing_targets"],
-        serde_json::json!([]),
-        "no targets declared so no targets missing"
-    );
-}
-
-#[test]
-fn doctor_reports_no_drift_when_everything_installed() {
-    let workspace = unique_temp_dir("doctor-no-drift");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         components = [\"clippy\"]\n",
-    );
-    let log_path = workspace.join("rustup.log");
-    // Rustup reports components as target-qualified — soldr doctor must
-    // treat `clippy-x86_64-...` as satisfying a declared `clippy`.
-    let rustup = install_doctor_fake_rustup(
-        &log_path,
-        &DoctorFakeRustupBehavior {
-            toolchain_list: vec!["1.94.1-x86_64-unknown-linux-gnu (default)".to_string()],
-            components_installed: vec!["clippy-x86_64-unknown-linux-gnu".to_string()],
-            targets_installed: vec!["x86_64-unknown-linux-gnu".to_string()],
-        },
-    );
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["doctor", "--json"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr doctor --json");
-
-    assert_eq!(
-        output.status.code(),
-        Some(0),
-        "doctor must exit 0 when no drift\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("doctor --json must produce JSON");
-    assert_eq!(json["drift"], false);
-    assert_eq!(json["missing_components"], serde_json::json!([]));
-    assert_eq!(json["missing_targets"], serde_json::json!([]));
-    let components = json["components"].as_array().expect("components array");
-    assert_eq!(components.len(), 1);
-    assert_eq!(components[0]["name"], "clippy");
-    assert_eq!(components[0]["installed"], true);
-}
-
-#[test]
-fn doctor_handles_missing_manifest() {
-    let workspace = unique_temp_dir("doctor-no-manifest");
-    let log_path = workspace.join("rustup.log");
-    // Failing fake rustup — if doctor invokes it the test fails because
-    // rustup will write to the log and exit non-zero.
-    let rustup = install_failing_fake_rustup(&log_path);
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["doctor"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr doctor");
-
-    assert_eq!(
-        output.status.code(),
-        Some(0),
-        "doctor must exit 0 when manifest is missing\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    assert!(
-        stdout.contains("no rust-toolchain.toml"),
-        "doctor stdout should mention missing manifest: {stdout}"
-    );
-
-    // The failing fake rustup writes to log_path on every invocation, so
-    // an empty (or non-existent) log proves rustup was never spawned.
-    let log = fs::read_to_string(&log_path).unwrap_or_default();
-    assert!(
-        log.trim().is_empty(),
-        "doctor must not invoke rustup when no manifest exists; log was: {log}"
-    );
-}
-
-#[test]
-fn doctor_reports_missing_target() {
-    let workspace = unique_temp_dir("doctor-missing-target");
-    seed_rust_toolchain_toml(
-        &workspace,
-        "[toolchain]\n\
-         channel = \"1.94.1\"\n\
-         targets = [\"x86_64-unknown-linux-musl\"]\n",
-    );
-    let log_path = workspace.join("rustup.log");
-    let rustup = install_doctor_fake_rustup(
-        &log_path,
-        &DoctorFakeRustupBehavior {
-            toolchain_list: vec!["1.94.1-x86_64-unknown-linux-gnu (default)".to_string()],
-            components_installed: Vec::new(),
-            // Only the host triple is installed — declared musl target
-            // is missing.
-            targets_installed: vec!["x86_64-unknown-linux-gnu".to_string()],
-        },
-    );
-
-    let output = Command::new(env!("CARGO_BIN_EXE_soldr"))
-        .args(["doctor", "--json"])
-        .current_dir(&workspace)
-        .env("SOLDR_TEST_RUSTUP_BIN", &rustup)
-        .output()
-        .expect("failed to run soldr doctor --json");
-
-    assert_eq!(
-        output.status.code(),
-        Some(1),
-        "doctor must exit 1 when target drift detected\nstdout:\n{}\nstderr:\n{}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let json: Value =
-        serde_json::from_slice(&output.stdout).expect("doctor --json must produce JSON");
-    assert_eq!(json["drift"], true);
-    assert_eq!(
-        json["missing_targets"],
-        serde_json::json!(["x86_64-unknown-linux-musl"])
-    );
-    assert_eq!(json["missing_components"], serde_json::json!([]));
-}
